% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajouter_ope_date.R
\name{ajouter_ope_date}
\alias{ajouter_ope_date}
\title{Rajouter la date de chaque opération à la passerelle.}
\usage{
ajouter_ope_date(passerelle, date_debut, date_fin = Sys.Date())
}
\arguments{
\item{passerelle}{Dataframe mettant en correspondance les identifiants des différentes tables}

\item{date_debut}{Date de début de la période à extraire au format jj/mm/aaaa}

\item{date_fin}{Date de fin de la période à extraire au format jj/mm/aaaa. Par défaut c'est la
date du jour où la fonction est exécutée.}
}
\value{
La passerelle à laquelle ont été ajoutés la date et l'année.
}
\description{
Rajouter la date de chaque opération à la passerelle.
}
\examples{
\dontrun{
passerelle <- ajouter_ope_date(passerelle = passerelle,
date_debut = '01/01/2020',
date_fin = '31/07/2020')
}
}
