% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspe-package.R
\docType{package}
\name{aspe-package}
\alias{aspe}
\alias{aspe-package}
\title{aspe: An R Package to Analyse and Visualise River Fish Data in France}
\description{
The aspe package provides a suite of tools for most of the common processing of the ASPE database \doi{10.1051/kmae/2022021}, including importation of the tables from a dump of the PostGreSQL database, calculation of abundances, densities, size distributions, along with graphical output and spatial processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PascalIrz/aspe/}
  \item \doi{10.5281/zenodo.7458005}
  \item \url{https://pascalirz.github.io/aspe/}
  \item Report bugs at \url{https://github.com/PascalIrz/aspe/issues}
}

}
\author{
\strong{Maintainer}: Pascal Irz \email{pascal.irz@ofb.gouv.fr} (\href{https://orcid.org/0000-0002-2066-8935}{ORCID})

Other contributors:
\itemize{
  \item Cédric Mondy \email{cedric.mondy@ofb.gouv.fr} (\href{https://orcid.org/0000-0003-2788-0936}{ORCID}) [contributor]
  \item Benoît Richard \email{benoit.richard@ofb.gouv.fr} (\href{https://orcid.org/0000-0003-4522-027X}{ORCID}) [contributor]
  \item Lilian Bonnafoux \email{l.bonnafoux@peche63.com} [contributor]
}

}
\keyword{internal}
