% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champs_aspe.R
\docType{data}
\name{champs_aspe}
\alias{champs_aspe}
\title{Signification des champs de la base Aspe}
\format{
An object of class \code{data.frame} with 718 rows and 3 columns.
}
\usage{
data(champs_aspe)
}
\description{
Signification des champs de la base Aspe
}
\examples{
\donttest{
data(champs_aspe)
champs_aspe}
}
\keyword{datasets}
