% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_coords_ligne.R
\name{convertir_coords_ligne}
\alias{convertir_coords_ligne}
\title{Convertir les coordonnées d'une ligne de dataframe dans un autre CRS}
\usage{
convertir_coords_ligne(
  ligne_liste,
  var_x,
  var_y,
  var_crs_initial,
  crs_sortie = 4326
)
}
\arguments{
\item{ligne_liste}{Une ligne d'un dataframe qui a été convertie en liste}

\item{var_x}{Caractère. Nom de la colonne qui contient la longitude.}

\item{var_y}{Caractère. Nom de la colonne qui contient la latitude.}

\item{var_crs_initial}{Caractère. Nom de la colonne qui contient le numéro EPSG du CRS de chacune
des observations (par exemple des stations ou des points).}

\item{crs_sortie}{Numérique entier. Le numéro EPSG du CRS de sortie. Par défaut
c'est en WGS84 (crs = 4326).}
}
\value{
Un dataframe à deux colonnes et une seule ligne contenant les coordonnées
    dans le CRS de sortie.
}
\description{
Convertir les coordonnées d'une ligne de dataframe dans un autre CRS
}
\examples{
\dontrun{
convertir_coords_ligne(ligne_liste = station[1], var_x = "sta_coordonnees_x",
    var_y = "sta_coordonnees_y", var_crs_initial = "typ_code_epsg", crs_sortie = 4326)
}

}
