% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creer_polygone_region.R
\name{creer_polygone_region}
\alias{creer_polygone_region}
\title{Créer un polygone pour uné région}
\usage{
creer_polygone_region(
  couche_shp_departements,
  departements_selectionnes,
  distance_buffer = 0.01,
  intitule_region
)
}
\arguments{
\item{couche_shp_departements}{Chemin vers la couche des départements au fromat .shp}

\item{departements_selectionnes}{Caractère. Vecteur des numéros des départements.}

\item{distance_buffer}{Numérique. Paramètre qui détermine la largeur du buffer.}

\item{intitule_region}{Caractère. Intitulé de la région.}
}
\value{
Le contour de la région (objet de classe sf).
}
\description{
Créer un polygone de classe sf à partir des numéros des départements et
    de la couche SIG des départements téléchargée
}
\examples{
\dontrun{
region <- creer_polygone_region (couche_shp_departements = url,
    departements_selectionnes = c("22", "29", "35", "56"),
    distance_buffer = 0.01,
    intitule_region = "Bretagne")
}
}
