% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionnaire.R
\docType{data}
\name{data_dictionnaire}
\alias{data_dictionnaire}
\title{Signification des champs de la base Aspe}
\format{
An object of class \code{data.frame} with 742 rows and 3 columns.
}
\usage{
data(data_dictionnaire)
}
\description{
Signification des champs de la base Aspe
}
\examples{
\donttest{
data(data_dictionnaire)
data_dictionnaire}
}
\keyword{datasets}
