% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_passerelle_taxo.R
\docType{data}
\name{data_passerelle_taxo}
\alias{data_passerelle_taxo}
\title{Table de passage entre les identifiants des espèces}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 266 rows and 5 columns.
}
\usage{
data(data_passerelle_taxo)
}
\description{
Tableau indiquant les correspondances entre les identifiants esp_id de la base Aspe,
asp_code_alternatif à trois lettres, taxref, sandre et le nom latin.
}
\examples{
\donttest{
data(data_passerelle_taxo)
data_passerelle_taxo}
}
\keyword{datasets}
