% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_protection_especes.R
\docType{data}
\name{data_protection_especes}
\alias{data_protection_especes}
\title{Données sur la protection réglementaire des espèces (poissons et écrevisses)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 2 columns.
}
\usage{
data(data_protection_especes)
}
\description{
Tableau avec une ligne par espèce protégée et une colonne indiquant le texte de référence.
}
\examples{
\donttest{
data("data_protection_especes")
data_protection_especes}
}
\keyword{datasets}
