% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_taille_poids.R
\docType{data}
\name{data_taille_poids}
\alias{data_taille_poids}
\title{Relations taille - poids}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 204 rows and 7 columns.
}
\usage{
data(data_taille_poids)
}
\description{
Tableau des relations taille calculées depuis la base ASPE ainsi que celles
issues de Fishbase. Pour ces dernières, les paramètres a et b ont été moyennés par espèce
en distinguant longueur fourche et longueur totale.
NB : Les unités sont les mêmes que dans la base de référence Fishbase,
à savoir pour la taille (en fait la longueur), le cm, et pour le poids (en fait
la masse), le gramme. Si l'on utilise les données de la base ASPE, il faut donc
opérer une conversion de longueur car celles-ci sont en mm.
}
\examples{
\donttest{
data(data_taille_poids)
data_taille_poids}
}
\keyword{datasets}
