% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_traits_bio.R
\docType{data}
\name{data_traits_bio}
\alias{data_traits_bio}
\title{Données sommaires sur les espèces de poissons}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124 rows and 22 columns.
}
\usage{
data(data_traits_bio)
}
\description{
Tableau d'une centaine de lignes correspondant aux espèces de poissons
référencées par leur code à trois lettres et de colonnes précisant le
statut de ces espèces ainsi que leur principaux habitats.
}
\examples{
\donttest{
data(data_traits_bio)
data_traits_bio}
}
\keyword{datasets}
