% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_attribut_id.R
\name{def_attribut_id}
\alias{def_attribut_id}
\title{Définir les attributs des points de prélèvement}
\usage{
def_attribut_id(df, var_id, var_tmp, var_pro = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{var_id}{Nom(s) de la ou des variable(s) contenant les identifiants des points de prélèvement (ex : pop_id, c(pop_id, saison))}

\item{var_tmp}{Nom de la variable contenant les dates de prélèvement (ex: annee)}

\item{var_pro}{[optionnelle] Nom de la variable contenant les protocoles de pêche}
}
\value{
Un dataframe avec le nombre d'observations (nb_obs) et le nombre de valeurs manquantes consecutives (nb_na_cons) pour chaque point de prélèvement ou, si \code{var_pro} est specifié, un dataframe avec le nombre d'observations (nb_obs), le nombre de valeurs manquantes consecutives (nb_na_cons), le nombre de protocoles de pêche (nb_pro) et le nombre de changements de protocole (nb_chg) pour chaque point de prélèvement et chaque combinaison de protocole de pêche.
}
\description{
Définie le nombre d'observations, de valeurs manquantes consecutives, de protocoles de pêche et de changements de protocole de chaque points de prélèvement
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
mef_ajouter_ope_date() \%>\%
mef_ajouter_type_protocole()
def_id <- def_attribut_id (var_id = pop_id,
var_tmp = annee,
var_pro = pro_libelle)
}
}
