% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionnaire.R
\docType{data}
\name{dictionnaire}
\alias{dictionnaire}
\title{Signification des champs de la base Aspe}
\format{
An object of class \code{data.frame} with 720 rows and 3 columns.
}
\usage{
data(dictionnaire)
}
\description{
Signification des champs de la base Aspe
}
\examples{
\donttest{
data(dictionnaire)
dictionnaire}
}
\keyword{datasets}
