% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_test.R
\docType{data}
\name{donneesTest}
\alias{donneesTest}
\alias{classe_ipr}
\alias{lot_poissons}
\alias{mesure_individuelle}
\alias{operation}
\alias{operation_description_peche}
\alias{operation_donnees_environnementales}
\alias{operation_ipr}
\alias{operation_objectif}
\alias{point_prelevement}
\alias{prelevement_elementaire}
\alias{ref_espece}
\alias{ref_logique_3}
\alias{ref_logique_4}
\alias{ref_moyen_prospection}
\alias{ref_objectif}
\alias{ref_protocole}
\alias{ref_type_longueur}
\alias{ref_type_lot}
\alias{ref_type_prelevement_elementaire}
\alias{ref_type_projection}
\alias{station}
\title{Extrait de la base Aspe utilisé pour les tests des fonctions et pour les vignettes}
\format{
An object of class \code{data.frame} with 7 rows and 9 columns.

An object of class \code{data.frame} with 19817 rows and 12 columns.

An object of class \code{data.frame} with 26062 rows and 10 columns.

An object of class \code{data.frame} with 55 rows and 46 columns.

An object of class \code{data.frame} with 55 rows and 29 columns.

An object of class \code{data.frame} with 55 rows and 26 columns.

An object of class \code{data.frame} with 55 rows and 42 columns.

An object of class \code{data.frame} with 73 rows and 2 columns.

An object of class \code{data.frame} with 8 rows and 42 columns.

An object of class \code{data.frame} with 101 rows and 4 columns.

An object of class \code{data.frame} with 245 rows and 12 columns.

An object of class \code{data.frame} with 3 rows and 3 columns.

An object of class \code{data.frame} with 4 rows and 3 columns.

An object of class \code{data.frame} with 4 rows and 3 columns.

An object of class \code{data.frame} with 16 rows and 5 columns.

An object of class \code{data.frame} with 7 rows and 5 columns.

An object of class \code{data.frame} with 3 rows and 4 columns.

An object of class \code{data.frame} with 4 rows and 5 columns.

An object of class \code{data.frame} with 3 rows and 3 columns.

An object of class \code{data.frame} with 32 rows and 6 columns.

An object of class \code{data.frame} with 8 rows and 16 columns.
}
\usage{
classe_ipr

lot_poissons

mesure_individuelle

operation

operation_description_peche

operation_donnees_environnementales

operation_ipr

operation_objectif

point_prelevement

prelevement_elementaire

ref_espece

ref_logique_3

ref_logique_4

ref_moyen_prospection

ref_objectif

ref_protocole

ref_type_longueur

ref_type_lot

ref_type_prelevement_elementaire

ref_type_projection

station
}
\description{
Il s'agit d'extraits des tables principales de la base qui ont été expurgés des taxons sensibles.
}
\details{
Ces tables sont décrites en détail dans https://doi.org/10.1051/kmae/2022021
}
\keyword{datasets}
