% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_tables_rdata.R
\name{exp_tables_rdata}
\alias{exp_tables_rdata}
\title{Exporter les dataframes au format .RData}
\usage{
exp_tables_rdata(repertoire = NA)
}
\arguments{
\item{repertoire}{Caractères. Chemin vers le répertoire où les fichiers seront écrits.}
}
\value{
Les fichiers sont sauvegardés dans le répertoire choisi.
}
\description{
Exporter les dataframes au format .RData
}
\examples{
\dontrun{
exp_tables_rdata()
}
}
