% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_commentaires_champs.R
\name{expl_commentaires_champs}
\alias{expl_commentaires_champs}
\title{Obtenir les commentaires qui donnent la signification des champs des tables}
\usage{
expl_commentaires_champs(fichier_dump)
}
\arguments{
\item{fichier_dump}{Le chemin vers le fichier d'extension .sql (décompressé) ou
.sql.gz (archive).}
}
\value{
Un dataframe avec pour chacun des champs sa table d'appartenance et sa signification.
}
\description{
Obtenir les commentaires qui donnent la signification des champs des tables
}
\examples{
\dontrun{
expl_commentaires_champs(fichier_dump = "../raw_data/dump.sql")
}


}
