% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_compter_champs.R
\name{expl_compter_champs}
\alias{expl_compter_champs}
\title{Sommer les champs contenus dans les différents dataframes d'un fichier .RData}
\usage{
expl_compter_champs(rdata)
}
\arguments{
\item{rdata}{Caractère. Chemin vers le fichier .RData}
}
\value{
La somme du nombre de champs dans chacun des dataframes
}
\description{
Sommer les champs contenus dans les différents dataframes d'un fichier .RData
}
\examples{
\dontrun{
expl_compter_champs('processed_data/toutes_tables_aspe_sauf_mei.RData')
[1] 748
}
}
