% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_lister_champs.R
\name{expl_lister_champs}
\alias{expl_lister_champs}
\title{Lister les champs contenus dans les différents dataframes d'un fichier .RData}
\usage{
expl_lister_champs(rdata)
}
\arguments{
\item{rdata}{Caractère. Chemin vers le fichier .RData}
}
\value{
Un dataframe avec pour chacun des champs (colonnes) son nom et celui de la table
qui le contient
}
\description{
Lister les champs contenus dans les différents dataframes d'un fichier .RData
}
\examples{
\dontrun{
synthese <- expl_lister_champs('processed_data/toutes_tables_aspe_sauf_mei.RData')
}
}
