% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_trouver_table.R
\name{expl_trouver_table}
\alias{expl_trouver_table}
\title{Rechercher les dataframes dont le nom contient un pattern}
\usage{
expl_trouver_table(nom_contient)
}
\arguments{
\item{nom_contient}{Chaine de caractères recherchés parmi les noms des tables.}
}
\value{
Un vecteur caractères contenant les noms des tables.
}
\description{
Rechercher les dataframes dont le nom contient un pattern
}
\examples{
\dontrun{
expl_trouver_table(nom_contient = "op")
}
}
