% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_trouver_variable.R
\name{expl_trouver_variable}
\alias{expl_trouver_variable}
\title{Rechercher les variables dont le nom contient un pattern}
\usage{
expl_trouver_variable(nom_contient)
}
\arguments{
\item{nom_contient}{Chaine de caractères recherchés parmi les noms des variables
des différentes tables.}
}
\value{
Un dataframe indiquant les noms de variables contenant la chaine recherchée
ainsi que leur table d'appartenance.
}
\description{
Rechercher les variables dont le nom contient un pattern
}
\examples{
\dontrun{
expl_trouver_variable(nom_contient = "ope")
}
}
