% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraire_ipr.R
\name{extraire_ipr}
\alias{extraire_ipr}
\title{Extraire les notes IPR à partir d'une table passerelle}
\usage{
extraire_ipr(passerelle, date_debut, date_fin = Sys.Date())
}
\arguments{
\item{passerelle}{Dataframe mettant en correspondance les identifiants des différentes tables}

\item{date_debut}{Date de début de la période à extraire au format jj/mm/aaaa}

\item{date_fin}{Date de fin de la période à extraire au format jj/mm/aaaa}
}
\value{
Dataframe avec la valeur de l'IPR, station, date + quelques informations complémentaires
}
\description{
Extraire les notes IPR à partir d'une table passerelle
}
\examples{
\dontrun{
mon_df_ipr <- extraire_ipr(passerelle = passerelle,
date_debut = '01/01/2020',
date_fin = '31/07/2020')
}

}
