% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraire_noms_tables.R
\name{extraire_noms_tables}
\alias{extraire_noms_tables}
\title{Extraire les noms des tables à partir des lignes du dump}
\usage{
extraire_noms_tables(lignes_dump)
}
\arguments{
\item{lignes_dump}{Nom de la liste qui contient les lignes du dump.}
}
\value{
Un vecteur caractères contenant les noms des tables.
}
\description{
Extraire les noms des tables à partir des lignes du dump
}
\examples{

\dontrun{
extraire_noms_tables (lignes_dump = mes_lignes_du_dump)
}


}
