% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_aggr_richesse.R
\name{geo_aggr_richesse}
\alias{geo_aggr_richesse}
\title{Aggréger les richesses par entité hydrographique}
\usage{
geo_aggr_richesse(
  ope_geo_data,
  echelle = c("Zone_Hydro", "Sous_Secteur_Hydro", "Secteur_Hydro", "Region_Hydro"),
  bassin_geo_poly
)
}
\arguments{
\item{ope_geo_data}{Objet de classe sf contenant les captures par opération.}

\item{echelle}{Caractère. Au choix entre "ZoneHydro", "SousSecteurHydro", "SecteurHydro"
et "RegionHydro".}

\item{bassin_geo_poly}{Objet de classe sf contenant les polygones délimitant les
entités hydrographiques.}
}
\value{
Un dataframe avec les richesses par entité hydrographique.
}
\description{
Calculer à partir des données Aspe le nombre d'espèces différentes recensées par
    Zone HYdro, Sous-secteur Hydro, Secteur Hydro ou Région hydro.
}
\examples{
\dontrun{
richesse_par_unite_spatiale <- geo_aggr_richesse(ope_geo_data = data_geo,
echelle = "RegionHydro",
bassin_geo_poly = bv_simp)
}
}
