% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_ajouter_coords_pop.R
\name{geo_ajouter_coords_pop}
\alias{geo_ajouter_coords_pop}
\title{Ajouter à la passerelle l'objectif de l'opération.}
\usage{
geo_ajouter_coords_pop(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Par objectif, on entent la contribution à un réseau (DCE, RCS, RRP, etc.), le suivi de poopulations,
    des études, du suivi de restauration, etc. Les modalités sont précisées dans la table ref_objectif.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
geo_ajouter_coords_pop()
}
}
