% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_ajouter_crs.R
\name{geo_ajouter_crs}
\alias{geo_ajouter_crs}
\title{Ajouter le code EPSG du CRS à un dataframe.}
\usage{
geo_ajouter_crs(df, var_id_crs)
}
\arguments{
\item{df}{Le dataframe de données.}

\item{var_id_crs}{Caractère. Le nom de la variable qui contient l'identifiant du CRS
au sens de la variable "typ_id" du dataframe "ref_type_projection" de la base Aspe.
selon la}
}
\value{
Le dataframe complété.
}
\description{
Ajouter le code EPSG du CRS à un dataframe.
}
\examples{
\dontrun{
pop <- point_prelevement \%>\%
geo_ajouter_crs(var_id_crs = pop_typ_id)
}
}
