% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_attr_dept_bv.R
\name{geo_attr_dept_bv}
\alias{geo_attr_dept_bv}
\title{Rajouter à l'objet sf des points de prélèvement les départements et (ou) les bassins auxquels
    ils appartiennent.}
\usage{
geo_attr_dept_bv(points_geo, depts_geo = NA, bv_geo = NA, crs_sortie = NA)
}
\arguments{
\item{points_geo}{Objet sf des points de prélèvement.}

\item{depts_geo}{Objet sf des polygones des départements.}

\item{bv_geo}{Objet sf des polygones des bassins versants (secteurs hydro ou autre).}

\item{crs_sortie}{Numérique. Identifiant EPSG du CRS de sortie. Par défaut, c'est celui de
points_geo.}
}
\value{
L'bjet sf des points de prélèvement complété des informations sur les départements
    et les bassins.
}
\description{
Rajouter à l'objet sf des points de prélèvement les départements et (ou) les bassins auxquels
    ils appartiennent.
}
\examples{
\dontrun{
points_geo <- points_geo \%>\%
geo_attr_dept_bv(depts_geo = shp_geoportail, bv_geo = secteurs_hydro_carthage)
}
}
