% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_attribuer.R
\name{geo_attribuer}
\alias{geo_attribuer}
\title{Rajouter à l'objet sf des points de prélèvement les départements et (ou) les bassins auxquels
    ils appartiennent.}
\usage{
geo_attribuer(points_geo, poly_sf, crs_sortie = NA)
}
\arguments{
\item{points_geo}{Objet sf des points qui se verront attribuer des poilygones.
points_geo doit avoir un CRS.}

\item{poly_sf}{Objet sf des polygones à attribuer (ex: départements, secteurs hydro).}

\item{crs_sortie}{Numérique. Identifiant EPSG du CRS de sortie. Par défaut, c'est celui de
points_geo.}
}
\value{
L'bjet sf des points de prélèvement complété des informations sur les départements
    et les bassins.
}
\description{
Rajouter à l'objet sf des points de prélèvement les départements et (ou) les bassins auxquels
    ils appartiennent.
}
\examples{
\dontrun{
points_geo <- points_geo \%>\%
geo_attribuer(poly_sf = secteurs_hydro_carthage)
}
}
