% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_completer_ref_type_projection.R
\name{geo_completer_ref_type_projection}
\alias{geo_completer_ref_type_projection}
\title{Compléter le référentiel des systèmes de coordonnées.}
\usage{
geo_completer_ref_type_projection(ref_type_projection)
}
\arguments{
\item{ref_type_projection}{Le dataframe contenant le référentiel des systèmes de coordonnées
de la base Aspe.}
}
\value{
Le dataframe contenant le référentiel des systèmes de coordonnées
    de la base Aspe, mais complété.
}
\description{
Dans la table `ref_type_projection`, pour le Lambert II étendu, le code EPSG (27572) est manquant.
    Cette fonction sert à le compléter.
}
\examples{
\dontrun{
ref_type_projection <- ref_type_projection \%>\%
geo_completer_ref_type_projection()
}
}
