% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_creer_polygone_bassins.R
\name{geo_creer_polygone_bassins}
\alias{geo_creer_polygone_bassins}
\title{Créer un objet sf avec polygones du découpage hydrographique à partir de geo.data.gouv.fr}
\usage{
geo_creer_polygone_bassins(url = NA, repertoire = "raw_data")
}
\arguments{
\item{url}{Texte. URL de téléchargement de la couche shapefile.}

\item{repertoire}{Texte. Chemin vers le répertoire de stockage et de décompression.}
}
\value{
Un objet sf multi polygones.
}
\description{
Créer un objet sf avec polygones du découpage hydrographique à partir de geo.data.gouv.fr
}
\examples{
\dontrun{
bv <- geo_creer_polygone_bassins(repertoire = "raw_data")
}
}
