% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ajouter_arriere_plan_int.R
\name{gg_ajouter_arriere_plan_int}
\alias{gg_ajouter_arriere_plan_int}
\title{Ajouter un arrière plan de classes à un ggplot}
\usage{
gg_ajouter_arriere_plan_int(graphique, df_classes)
}
\arguments{
\item{graphique}{Graphique ggplot.}

\item{df_classes}{Dataframe contenant les limites de classes comme classe_ipr.}
}
\value{
Le graphique complété par son arrière-plan coloré.
}
\description{
Ajouter un arrière plan de classes à un ggplot
}
\examples{
\dontrun{
ggplot(data = data) \%>\%
  gg_ajouter_arriere_plan_int()
}
}
