% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_colo_ext_pop.R
\name{gg_colo_ext_pop}
\alias{gg_colo_ext_pop}
\title{Produire le graphique de la dynamique du peuplement en un point de prélèvement.}
\usage{
gg_colo_ext_pop(df)
}
\arguments{
\item{df}{Le dataframe contenant les données, mises en forme par la fonction mef_colo_ext_pop().}
}
\value{
Le graphique produit avec ggplot2.
}
\description{
Produire le graphique de la dynamique du peuplement en un point de prélèvement.
}
\examples{
\dontrun{
gg_colo_ext_pop(df = colo_ext_ma_station)
}
}
