% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_colo_ext_pops.R
\name{gg_colo_ext_pops}
\alias{gg_colo_ext_pops}
\title{Produire le graphique de la dynamique du peuplement pour chacun des points de prélèvement
    du dataframe.}
\usage{
gg_colo_ext_pops(df)
}
\arguments{
\item{df}{Le dataframe contenant les données, mises en forme par la fonction mef_colo_ext_pops().}
}
\value{
Une liste contenant les graphiques produits avec ggplot2.
}
\description{
Produire le graphique de la dynamique du peuplement pour chacun des points de prélèvement
    du dataframe.
}
\examples{
\dontrun{
mes_graphiques <- gg_colo_ext_pops(df = mon_df)
mes_graphiques[[1]]
}
}
