% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_dyn_esp_pop.R
\name{gg_dyn_esp_pop}
\alias{gg_dyn_esp_pop}
\title{Produire un graphique de synthèse de la dynamique d'une population}
\usage{
gg_dyn_esp_pop(df, libelle_point, nom_espece, seuil = 0)
}
\arguments{
\item{df}{Dataframe contenant les données, avec les variables "pop_id", "pop_libelle",
esp_nom_commun", "mei_taille" et "annee".}

\item{libelle_point}{Caractère. Nom du point dans le champ "pop_libelle".}

\item{nom_espece}{Caractère. Nom de l'espèce dans le champ "esp_nom_commun".}

\item{seuil}{Numérique entre 0 et 1. Seuil de pourcentile pour exclure des histogrammes les rares
individus de longueur extrême. Par défaut fixé à 0 donc toutes les observations sont conservées.
S'il est fixé à 0.01, les 1% des longueurs les plus élevées sont exclus des histogrammes. Ces
individus sont en revanche conservés sur les autres graphiques.}
}
\value{
Le graphique ggplot2 avec les distributions en taille par année et l'évolution
    temporelle de la longueur médiane et des effectifs capturés.
}
\description{
Représentation pour une espèce sur un point de prélèvement.
}
\examples{
\dontrun{
gg_dyn_esp_pop(df = data,
               libelle_point = "DROME A CHABRILLAN ",
               esp_nom = "Barbeau fluviatile")
}
}
