% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_env_fn_annee.R
\name{gg_env_fn_annee}
\alias{gg_env_fn_annee}
\title{Produire le graphique de l'évolution temporelle d'un facteur environnemental moyen comparé
    entre les sites de présence et d'absence des espèces.}
\usage{
gg_env_fn_annee(df, parametre, annee_debut = 1995, annee_fin = NA)
}
\arguments{
\item{df}{Le dataframe contenant les données avec les variables presence, parametre,
valeur_parametre, annee}

\item{parametre}{Le paramètre (ex : "altitude") représenté.}

\item{annee_debut}{Année mini à représenter.}

\item{annee_fin}{Année maxi à représenter.}
}
\value{
Le graphique ggplot2.
}
\description{
Produire le graphique de l'évolution temporelle d'un facteur environnemental moyen comparé
    entre les sites de présence et d'absence des espèces.
}
