% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_gerer_seuils_classes_ipr_int.R
\name{gg_gerer_seuils_classes_ipr_int}
\alias{gg_gerer_seuils_classes_ipr_int}
\title{Gérer les limites de classes d'indices sur classe_ipr}
\usage{
gg_gerer_seuils_classes_ipr_int(df_classes, metriques = FALSE, sup_500m)
}
\arguments{
\item{df_classes}{Dataframe d'entrée, en général classe_ipr}

\item{metriques}{Booléen. Si TRUE, on considère qu'il s'agit des métriques et non de
l'indice agrégé. Les seuils sont donc divisés par 7 (le nombre des métriques dans l'IPR).}

\item{sup_500m}{Booléen. TRUE si les stations sont au-dessus de 500m d'altitude, ce qui change
un seuil de classe IPR.}
}
\value{
Le dataframe filtré selon l'altitude
}
\description{
Gérer les limites de classes d'indices sur classe_ipr
}
\examples{
\dontrun{
classe_classes_ipr <- classe_ipr \%>\%
  gg_gerer_seuils_classes_ipr_int()
}
}
