% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ipr_metrique.R
\name{gg_ipr_metrique}
\alias{gg_ipr_metrique}
\title{Produire le barplot des métriques IPR en fonction de l'année sur une station}
\usage{
gg_ipr_metrique(
  df,
  var_station = "sta_id",
  station,
  var_metrique,
  var_valeur_metrique,
  metriques_selection = "opi_ipr",
  classes = "IPR",
  palette_couleurs = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E",
    "#E6AB02", "#A6761D", "#666666")
)
}
\arguments{
\item{df}{Le dataframe contenant les données nécessaires, dont les variables désignées par
les arguments ci-dessous var_ ainsi que la variable **annee**.}

\item{var_station}{Caractère. Nom de la variable qui contient l'identifiant de la station.
Par défaut prend la valeur "sta_id" pour une identification par ce champ dans la table
"station" de la base.}

\item{station}{Caractère. Identifiant de la station.}

\item{var_metrique}{Caractère. Nom de la variable qui contient l'identifiant des métriques.}

\item{var_valeur_metrique}{Caractère. Nom de la variable qui contient les valeurs des métriques.}

\item{metriques_selection}{Vecteur caractères. Noms des métriques à représenter.
Par défaut prend la valeur "opi_ipr" pour le champ IPR ainsi nommé dans la table "operation_ipr".}

\item{classes}{Caractère. Peut prendre les valeurs "IPR" (par défaut) ou "metriques". Pour l'IPR les
seuils des classes de qualité sont 0, 5, 16, 25, 36. Comme l'IPR est la somme des 7 métriques, les seuils
sont divisés par 7 si l'on représente les métriques individuelles.}

\item{palette_couleurs}{Vecteur caractères. Palette de 8 couleurs pour les 7 métriques + l'IPR agrégé.
Par défaut empruntée à RColorBrewer::brewer.pal(n = 8, name = "Dark2").}
}
\value{
La graphique ggplot2.
}
\description{
Produire le barplot des métriques IPR en fonction de l'année sur une station
}
\examples{
\dontrun{
df \%>\%
gg_ipr_metrique(
  var_station = "Sta_Ann",
  station = "Barr",
  var_metrique = "IPR_metriq",
  var_valeur_metrique = "valeur",
  metriques_selection = c("NER", "NEL", "NTE"),
  classes = "metriques"
  )
}
}
