% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ipr_station.R
\name{gg_ipr_station}
\alias{gg_ipr_station}
\title{Graphique de la série chronologique des IPR}
\usage{
gg_ipr_station(df_ipr, var_id_sta, station_sel, sup_500m = FALSE)
}
\arguments{
\item{df_ipr}{Dataframe contenant les données IPR.}

\item{var_id_sta}{Nom de la variable servant à identifier les stations ou points.
Cette variable donnera les étiquettes du graphique.}

\item{station_sel}{Vecteur caractère indiquant les points ou stations à sélectionner.}

\item{sup_500m}{Booléen. Indique si les stations sont situées à des altitudes inférieures
(sup_500m = FALSE, par défaut) ou bien supérieures à 500m car les seuils de classe
se qualité varient selon l'altitude.}
}
\value{
Un graphique ggplot2.
}
\description{
La fonction utilise le référentiel "classe_ipr" de la base Aspe qui doit donc être
    chargé auparavant et complété par les codes couleurs avec ipr_completer_classes_couleur().
}
\examples{
\dontrun{
classe_ipr <- classe_ipr \%>\%
  ipr_completer_classes_couleur()
}
}
