% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_temp_indic_div.R
\name{gg_temp_indic_div}
\alias{gg_temp_indic_div}
\title{Graphique de la série chronologique de métriques de diversité à l'échelle du peuplement sur une station ou un groupe de stations}
\usage{
gg_temp_indic_div(
  df_captures,
  var_id_sta,
  var_metrique_div,
  station_sel = NULL,
  nb_colonnes = 2,
  id_sta_max_caract = 25,
  print.metrique = F
)
}
\arguments{
\item{df_captures}{Dataframe contenant les données des effectifs capturés par taxons. Il doit contenir les variables "effectif" et "annee", ainsi que
trois variables pour d'identifier les stations (ou points de prélèvement) pour les calculs.}

\item{var_id_sta}{Variable servant à identifier les stations ou points.
Cette variable donnera les étiquettes des lignes du graphique.}

\item{var_metrique_div}{Variable donnant le nom de la colonne des métriques (ex : metriques_div).}

\item{station_sel}{Vecteur caractère indiquant les points ou stations à sélectionner.}

\item{nb_colonnes}{Entier. Nombre de colonnes du graphique utilisé pour dans le cas d'un graphique pour une station. Par défaut nb_colonnes = 2.}

\item{id_sta_max_caract}{Entier. Nombre maximum de caractères dans l'identifiant de la station, au-delà duquel il sera
découpé pour tenir sur plusieurs lignes. Par défaut c'est 25 caractères.}

\item{print.metrique}{Valeur logique. Si TRUE, le dataframe des métriques est retourné. Si FALSE (par défaut), le graphique \code{ggplot2} est retourné.}
}
\value{
Un graphique \code{ggplot2} ou un dataframe.
}
\description{
Cette fonction peut renvoyer soit:
  (i) un dataframe avec les métriques calculées (Richesse taxonomique, Effectif total, Indices de Shannon et Equitabilité),
  (ii) ou à un graphique \code{ggplot2} en treillis avec les différentes métriques, pour une station ou un groupe de stations.
}
\examples{
\dontrun{
# preparation donnees
captures <-
  mef_creer_passerelle() \%>\% # création passerelle
  mef_ajouter_ope_date() \%>\%  # ajout date et année
  mef_ajouter_lots() \%>\%  # ajout des lots (effectif par espèce)
  mef_ajouter_surf_calc() \%>\%  # ajout surfaces prospectées pour calcul des densités
  mef_ajouter_libelle() \%>\% # ajout libellé du point de prélèvement
  group_by(
    sta_id,
    pop_id,
    ope_id,
    ope_date,
    annee,
    esp_code_alternatif,
    ope_surface_calculee,
    pop_libelle
  ) \%>\%
  summarise(effectif = sum(lop_effectif, na.rm = TRUE)) \%>\%
  ungroup() \%>\%
  mef_ajouter_type_protocole() \%>\%
  filter(pro_libelle \%in\% c("Pêche complète à un ou plusieurs passages",
                            "Pêche partielle par points (grand milieu)",
                            "Pêche par ambiances",
                            "Pêche partielle sur berge"))
# graphiques ggplot
gg_temp_indic_div(df_captures = captures,
                    var_id_sta = pop_libelle,
                    var_metrique_div = metrique_div,
                    station_sel =c("LA BERENCE A GAVRAY", "La Berre à Portel-des-Corbières"),
                    print.metrique = F)

gg_temp_indic_div(df_captures = captures,
                    var_id_sta = pop_libelle,
                    var_metrique_div = metrique_div,
                    station_sel =c("La Berre à Portel-des-Corbières"),
                    nb_colonnes = 2,
                    print.metrique = F)

# dataframe des metriques
gg_temp_indic_div(df_captures = captures,
                    var_id_sta = pop_libelle,
                    var_metrique_div = metrique_div,
                    station_sel =c("LA BERENCE A GAVRAY", "La Berre à Portel-des-Corbières"),
                    print.metrique = T)

gg_temp_indic_div(df_captures = captures,
                    var_id_sta = pop_libelle,
                    var_metrique_div = metrique_div,
                    station_sel =c("La Berre à Portel-des-Corbières"),
                    nb_colonnes = 2,
                    print.metrique = T)
}
}
