% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_temp_ipr_pc_bon.R
\name{gg_temp_ipr_pc_bon}
\alias{gg_temp_ipr_pc_bon}
\alias{ipr_grapher_pc_bon}
\title{Produire un graphique du pourcentage de stations en bon état au plan IPR sur un jeu de
stations, en fonction de l'année}
\usage{
gg_temp_ipr_pc_bon(ipr_df, titre = NA)
}
\arguments{
\item{ipr_df}{Dataframe contenant les données, issu de la fonction extraire_ipr().}

\item{titre}{Caractère. Titre du graphique.}
}
\value{
Le graphique issu de ggplot2.
}
\description{
Produire un graphique du pourcentage de stations en bon état au plan IPR sur un jeu de
stations, en fonction de l'année
}
\examples{
\dontrun{
gg_temp_ipr_pc_bon(ipr_df = data_56,
titre = "Morbihan")
}
}
