% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_temp_metriq_grille.R
\name{gg_temp_metriq_grille}
\alias{gg_temp_metriq_grille}
\title{Graphique de la série chronologique des métriques IPR sur une station ou un groupe de stations}
\usage{
gg_temp_metriq_grille(
  df_metriques,
  var_id_sta,
  var_nom_metrique,
  var_valeur_metrique,
  station_sel = NA,
  max_axe_y = 10,
  inv_y = TRUE
)
}
\arguments{
\item{df_metriques}{Dataframe contenant les données. Il doit contenir une variable "annee" ainsi que
trois variables pour d'identifier les stations (ou points de prélèvement) et les métriques, ainsi que
pour donner la valeur des métriques.}

\item{var_id_sta}{Variable servant à identifier les stations ou points.
Cette variable donnera les étiquettes des lignes du graphique.}

\item{var_nom_metrique}{Variable contenant les noms des métriques (ex : dio, dti).}

\item{var_valeur_metrique}{Variable numérique contenant les valeurs des métriques.}

\item{station_sel}{Vecteur caractère indiquant les points ou stations à sélectionner.}

\item{max_axe_y}{Numérique. Limite supérieure de l'axe des ordonnées. Par défaut max_axe_y = 10.}

\item{inv_y}{Booléen. Indique l'axe des ordonnées pointe vers le bas (TRUE, par défaut) ou
vers le haut. NB pour l'IPR, plus l'indice est faible plus la qualité est élevée.
C'est l'inverse pour l'IPR+.}
}
\value{
Un graphique ggplot2.
}
\description{
Graphique de la série chronologique des métriques IPR sur une station ou un groupe de stations
}
\examples{
\dontrun{
# tidy a data set
metriques <- mef_creer_passerelle() \%>\%
select(-lop_id, -pre_id) \%>\%
  distinct() \%>\%
  mef_ajouter_metriques() \%>\%
  mef_ajouter_libelle() \%>\%
  mef_ajouter_ope_date() \%>\%
  filter(!is.na(ner)) \%>\%
  select(-ends_with("observe"), -ends_with("theorique")) \%>\%
  pivot_longer(cols = ner:dti,
               names_to = "metrique",
               values_to = "valeur")

# the plot
gg_temp_metriq_grille(df_metriques = metriques,
                      station_sel = c("La Berre à Portel-des-Corbières", "LA BERENCE A GAVRAY"),
                      var_id_sta = pop_libelle,
                      var_nom_metrique = metrique,
                      var_valeur_metrique = valeur)
}
}
