% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-connexion-idg.R
\name{imp_connecter_aspe_idg}
\alias{imp_connecter_aspe_idg}
\title{Démarre une connexion à ASPE sur la BDD IDG (OFB)}
\usage{
imp_connecter_aspe_idg()
}
\value{
Connexion. Renvoie une erreur si connexion impossible.
}
\description{
Fonction pour se connecter à la BDD ASPE en utilisant les paramètres de connexion
stockés dans l'environnement R (fichier \code{.Renviron}).
}
\examples{
\dontrun{
bdd_aspe <- imp_connecter_aspe_idg()

}

}
