% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_corres_aspe_taxref.R
\name{imp_corres_aspe_taxref}
\alias{imp_corres_aspe_taxref}
\title{Importer une table de correspondance entre les codes espèces 3 lettres Aspe et leurs codes Taxref}
\usage{
imp_corres_aspe_taxref(
  url = "https://api.sandre.eaufrance.fr/referentiels/v1/apt.csv?compress=true"
)
}
\arguments{
\item{url}{Caractère. URL du jeu de données à télécharger.
Par défaut "https://api.sandre.eaufrance.fr/referentiels/v1/apt.csv?compress=true"}
}
\value{
Dataframe. Table de correspondance
}
\description{
La donnée ne provient pas de la base ASPE, mais de l'API SANDRE.
}
