% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-connexion-idg.R
\name{imp_deconnecter_aspe_idg}
\alias{imp_deconnecter_aspe_idg}
\title{Stoppe une connexion à ASPE sur la BDD IDG (OFB)}
\usage{
imp_deconnecter_aspe_idg(conn)
}
\arguments{
\item{conn}{nom de la connexion (objet) à stopper.}
}
\value{
Aucune valeur.
}
\description{
Fonction stoppe la connexion à la BDD
}
\examples{
\dontrun{
bdd_aspe <- imp_connecter_aspe_idg()

imp_deconnecter_aspe_idg(conn = bdd_aspe)
}
}
