% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_extraire_date_fichier.R
\name{imp_extraire_date_fichier}
\alias{imp_extraire_date_fichier}
\title{Obtenir la date du dump SQL de la base Aspe}
\usage{
imp_extraire_date_fichier(fichier, date = "modification")
}
\arguments{
\item{fichier}{Le chemin vers le fichier.}

\item{date}{Date souhaitée, soit "creation", soit "modification" (par défaut).}
}
\value{
La date de création ou de modification.
}
\description{
Obtenir la date du dump SQL de la base Aspe
}
\examples{
\dontrun{
date-dump <- imp_extraire_date_fichier(fichier = "../raw_data/dump.sql",
date = "creation")
}


}
