% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_extraire_noms_tables2.R
\name{imp_extraire_noms_tables2}
\alias{imp_extraire_noms_tables2}
\title{Extraire les noms des tables à partir des lignes du dump}
\usage{
imp_extraire_noms_tables2(lignes_dump)
}
\arguments{
\item{lignes_dump}{Nom de la liste qui contient les lignes du dump.}
}
\value{
Un vecteur caractères contenant les noms des tables.
}
\description{
Extraire les noms des tables à partir des lignes du dump
}
\examples{

\dontrun{
imp_extraire_noms_tables (lignes_dump = mes_lignes_du_dump)
}


}
