% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_lire_lignes_dump.R
\name{imp_lire_lignes_dump}
\alias{imp_lire_lignes_dump}
\title{Lire les lignes du dump SQL de la base Aspe}
\usage{
imp_lire_lignes_dump(fichier_dump)
}
\arguments{
\item{fichier_dump}{Le chemin vers le fichier d'extension .sql (décompressé) ou
.sql.gz (archive).}
}
\value{
Une liste de chaînes de caractères avec autant d'éléments que de lignes sur le dump.
}
\description{
Lire les lignes du dump SQL de la base Aspe
}
\examples{

\dontrun{
lignes_dump <- imp_lire_lignes_dump (fichier_dump = "../raw_data/dump.sql")
}


}
