% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_scinder_dump.R
\name{imp_scinder_dump}
\alias{imp_scinder_dump}
\title{Scinder le dump en deux parties pour les traiter séparément et éviter de
saturer la mémoire vive}
\usage{
imp_scinder_dump(fichier_dump, repertoire_sortie = NA)
}
\arguments{
\item{fichier_dump}{Le chemin vers le fichier d'extension .sql (décompressé) ou
.sql.gz (archive).}

\item{repertoire_sortie}{Le chemin vers le répertoire de sortie (qui doit avoir
été créé auparavant). Par défaut c'est celui qui contient le dump.}
}
\value{
Deux fichiers au format .RData contenant l'un les lignes de la table
mesure_individuelle et l'autre tout le reste du dump
}
\description{
ATTENTION : Si le chemin d'accès au fichier est trop long, il est possible que
la fonction retourne une erreur :
Error in gzfile(file, "wb") : argument 'description' incorrect
In if (!nzchar(file)) stop("'file' must be non-empty string") :
a condition a une longueur > 1 et seul le premier élément est utilisé
Dans ce cas, déplacez le dump pour accourcir le chemin d'accès.
}
\examples{
\dontrun{
imp_scinder_dump(fichier_dump = "../raw_data/dump.sql")
}

}
