% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_tables_a_partir_des_lignes2.R
\name{imp_tables_a_partir_des_lignes2}
\alias{imp_tables_a_partir_des_lignes2}
\title{Importer des tables à partir des lignes du dump de la base Aspe}
\usage{
imp_tables_a_partir_des_lignes2(lignes_dump, tables_a_extraire = NA)
}
\arguments{
\item{lignes_dump}{Nom de la liste contenant les lignes du dump.}

\item{tables_a_extraire}{Un vecteur contenant les noms des tables à extraire.
Par défaut toutes les tables sont extraites, ce qui peut prendre du temps
inutilement.}
}
\value{
Les dataframes correspondant à chacune des tables de la base.
}
\description{
Importer des tables à partir des lignes du dump de la base Aspe
}
\examples{
\dontrun{
imp_tables_a_partir_des_lignes(lignes_dump = lignes_autres_tables)
}


}
