% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_trouver_index_fin.R
\name{imp_trouver_index_fin}
\alias{imp_trouver_index_fin}
\title{Trouver la ligne de fin qui correspond à une table désignée par sa ligne de début}
\usage{
imp_trouver_index_fin(vecteur_index_lignes_fin, ligne_debut)
}
\arguments{
\item{vecteur_index_lignes_fin}{Vecteur contenant les numéros de toutes les lignes contenant la séquence caractéristique
de la fin d'une table.}

\item{ligne_debut}{Numéro de la ligne de début de la table.}
}
\value{
Le numéro de la ligne de fin de la table.
}
\description{
Par exemple pour trouver le numéro de la ligne de fin d'une table qui débuterait à la ligne 6517552
}
\examples{
\dontrun{
imp_trouver_index_fin(vecteur_index_lignes_fin = mon_vecteur, ligne_debut = 6517552)
}
}
