% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_tables_a_partir_des_lignes.R
\name{importer_tables_a_partir_des_lignes}
\alias{importer_tables_a_partir_des_lignes}
\title{Importer des tables à partir des lignes du dump de la base Aspe}
\usage{
importer_tables_a_partir_des_lignes(lignes_dump, tables_a_extraire = NA)
}
\arguments{
\item{lignes_dump}{Nom de la liste contenant les lignes du dump.}

\item{tables_a_extraire}{Un vecteur contenant les noms des tables à extraire.
Par défaut toutes les tables sont extraites, ce qu ipeut prendre du temps
unutilement.}
}
\value{
Les dataframes correspondant à chacune des tables de la base.
}
\description{
Importer des tables à partir des lignes du dump de la base Aspe
}
\examples{
\dontrun{
importer_dump_sql(fichier_dump = "../raw_data/dump.sql")
}


}
