% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_completer_classes_couleur.R
\name{ip_completer_classes_couleur}
\alias{ip_completer_classes_couleur}
\title{Ajouter les codes couleurs aux classes IPR}
\usage{
ip_completer_classes_couleur(
  df_classes_ip,
  var_numero_classe = cli_classe,
  var_libelle_classe = cli_libelle
)
}
\arguments{
\item{df_classes_ip}{Dataframe contenant les intitulés des classes, typiquement
le dataframe classe_ipr de la base Aspe.}

\item{var_numero_classe}{Variable contenant les numéros de classes de 1 (meilleur état)
à 5 (le pire).}

\item{var_libelle_classe}{Variable contenant les libellés de classes.}
}
\value{
Le dataframe complété. La colonne ajoutée est "classe_couleur".
}
\description{
Ajouter les codes couleurs aux classes IPR
}
\examples{
\dontrun{
classe_ipr_plus <- classe_ipr_plus \%>\%
  ip_completer_classes_couleur(var_numero_classe = cip_classe,
                               var_libelle_classe = cip_libelle)
}
}
