% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr_completer_classes_couleur.R
\name{ipr_completer_classes_couleur}
\alias{ipr_completer_classes_couleur}
\title{Ajouter les codes couleurs aux classes IPR}
\usage{
ipr_completer_classes_couleur(df_classes_ipr)
}
\arguments{
\item{df_classes_ipr}{Dataframe contenant les intitulés des classes, typiquement
le dataframe classe_ipr de la base Aspe.}
}
\value{
Le dataframe complété. La colonne ajoutée est "cli_couleur".
}
\description{
Ajouter les codes couleurs aux classes IPR
}
\examples{
\dontrun{
gg_ipr_station(df_ipr = ipr_dept,
  var_id_sta = pop_libelle,
  station_sel = c("TRIEUX à PLESIDY", "LEGUER à PLOUBEZRE"),
  sup_500m = FALSE)
}
}
