% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr_extraire.R
\name{ipr_extraire}
\alias{ipr_extraire}
\title{Extraire les notes IPR à partir d'une table passerelle}
\usage{
ipr_extraire(
  passerelle,
  date_debut = NA,
  date_fin = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{passerelle}{Dataframe mettant en correspondance les identifiants des différentes tables.}

\item{date_debut}{Date de début de la période à extraire au format jj/mm/aaaa.}

\item{date_fin}{Date de fin de la période à extraire au format jj/mm/aaaa. Par défaut c'est la
date du jour où la fonction est exécutée.}
}
\value{
Dataframe avec la valeur de l'IPR, station, date + quelques informations complémentaires
}
\description{
Extraire les notes IPR à partir d'une table passerelle
}
\examples{
\dontrun{
mon_df_ipr <- ipr_extraire(passerelle = passerelle,
date_debut = '01/01/2020',
date_fin = '31/07/2020')
}

}
