% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr_extraire_metriques.R
\name{ipr_extraire_metriques}
\alias{ipr_extraire_metriques}
\title{Extraire les métriques individuelles observées, théoriques et leur contribution
    à l'indice IPR agrégé.}
\usage{
ipr_extraire_metriques(passerelle)
}
\arguments{
\item{passerelle}{Le dataframe passerelle.}
}
\value{
Le dataframe passerelle complété par les métriques.
}
\description{
Extraire les métriques individuelles observées, théoriques et leur contribution
    à l'indice IPR agrégé.
}
\examples{
\dontrun{
metriques <- ipr_extraire_metriques(passerelle = ma_passerelle)
}
}
