% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr_formater_pour_macro.R
\name{ipr_formater_pour_macro}
\alias{ipr_formater_pour_macro}
\title{Mettre les données nécessaires au calcul de l'IPR au format requis en entrée de la macro
Excel de calcul de l'indice}
\usage{
ipr_formater_pour_macro(
  passerelle,
  date_debut,
  date_fin = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{passerelle}{Dataframe mettant en correspondance les identifiants des différentes tables.}

\item{date_debut}{Date de début de la période à extraire au format jj/mm/aaaa.}

\item{date_fin}{Date de fin de la période à extraire au format jj/mm/aaaa. Par défaut c'est la
date du jour où la fonction est exécutée.}
}
\value{
Un dataframe au format souhaité, mais dont les colonnes doivent ensuite être renommées
avec la fonction renommer_pour_macro().
}
\description{
Mettre les données nécessaires au calcul de l'IPR au format requis en entrée de la macro
Excel de calcul de l'indice
}
\examples{
\dontrun{
data_macro <- ipr_formater_pour_macro(passerelle = passerelle, date_debut = '01/01/2020')
}
}
