% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr_mef_radar.R
\name{ipr_mef_radar}
\alias{ipr_mef_radar}
\title{Mettre un dataframe au format pour ls graphiques en radar.
  La fonction radarchart() du package {fmsb} nécessite en entrée un dataframe
  au format spécifique.}
\usage{
ipr_mef_radar(metriques_df, station_id)
}
\arguments{
\item{metriques_df}{Un dataframe contenant les métriques par station et par année,
produit par la fonction ipr_extraire_metriques().}

\item{station_id}{Numérique entier. Identifiant de la station (sta_id).}
}
\value{
Un dataframe mis en forme.
}
\description{
Mettre un dataframe au format pour ls graphiques en radar.
  La fonction radarchart() du package {fmsb} nécessite en entrée un dataframe
  au format spécifique.
}
\examples{
\dontrun{
ipr_mef_radar <- ipr_mef_radar(metriques_df = metriques_22, station_id = 10708)
}
}
