% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr_pivoter_1colonne_par_an.R
\name{ipr_pivoter_1colonne_par_an}
\alias{ipr_pivoter_1colonne_par_an}
\title{Convertir une table des IPR du format long au format large}
\usage{
ipr_pivoter_1colonne_par_an(ipr_df)
}
\arguments{
\item{ipr_df}{Le dataframe au format long contenant les IPR, issi de la fonction extraire-ipr()}
}
\value{
Un dataframe contenant les mêmes données, mais agencé avec une colonne par année.
}
\description{
On appelle format large la présentation des données avec une colonne par année.
S'il y a plusieurs valeurs de l'IPR pour un même point une année donnée,
la valeur retournée sera la moyenne.
}
\examples{
\dontrun{
ipr_large_df <- ipr_pivoter_1colonne_par_an(ipr_df = mon_df_ipr)
}
}
