% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lire_lignes_dump_sql.R
\name{lire_lignes_dump_sql}
\alias{lire_lignes_dump_sql}
\title{Lire les lignes du dump SQL de la base Aspe}
\usage{
lire_lignes_dump_sql(fichier_dump)
}
\arguments{
\item{fichier_dump}{Le chemin vers le fichier d'extension .sql (décompressé) ou
.sql.gz (archive).}
}
\value{
Une liste de chaînes de caractères avec autant d'éléments que de lignes sur le dump.
}
\description{
Lire les lignes du dump SQL de la base Aspe
}
\examples{

\dontrun{
lignes_dump <- lire_lignes_dump_sql (fichier_dump = "../raw_data/dump.sql")
}


}
