% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_liste_rouge.R
\docType{data}
\name{liste_rouge}
\alias{liste_rouge}
\title{Données sur le statut liste rouge des espèces de poissons}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 110 rows and 22 columns.
}
\usage{
data(liste_rouge)
}
\description{
Tableau avec en lignes desespèces de poissons et en colonnes les données sur le
statut liste rouge de ces espèces.
}
\examples{
\donttest{
data(liste_rouge)
liste_rouge}
}
\keyword{datasets}
