% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_esp.R
\name{mef_ajouter_esp}
\alias{mef_ajouter_esp}
\title{Ajouter à un df "passerelle le code espèce à 3 lettres, son nom commun et scientifique}
\usage{
mef_ajouter_esp(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
La passerelle doit comprendre un champ avec le code taxon esp_id (typiquement, on utilise donc
mef_ajouter_esp() après mef_ajouter_lots())}
}
\value{
La passerelle complétée.
}
\description{
Les codes sont "historiques", hérités du Conseil Supérieur de la Pêche.
}
\examples{
\dontrun{
passerelle <- mef_creer_passerelle() \%>\%
mef_ajouter_lots() \%>\%
mef_ajouter_esp()
}
}
