% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_esp_code_alternatif.R
\name{mef_ajouter_esp_code_alternatif}
\alias{mef_ajouter_esp_code_alternatif}
\title{Ajouter le code espèce à 3 lettres à la passerelle.}
\usage{
mef_ajouter_esp_code_alternatif(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter le code espèce à 3 lettres à la passerelle.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_esp_code_alternatif()
}
}
