% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_groupe_points.R
\name{mef_ajouter_groupe_points}
\alias{mef_ajouter_groupe_points}
\title{Ajouter à la passerelle, quand le prélèvement élémentaire est un groupe de points, ses données}
\usage{
mef_ajouter_groupe_points(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Le type de prélèvement doit donc être indiqué dans le data frame (Groupe de points,
Passage ou Ambiance). On applique mef_ajouter_groupe_points() après mef_ajouter_type_prelevement().
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_type_prelevement() \%>\%
mef_ajouter_groupe_points()
}
}
