% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_intervenants.R
\name{mef_ajouter_intervenants}
\alias{mef_ajouter_intervenants}
\title{Ajouter à la passerelle les intervenants sur les opérations
    Ces intervenants sont l'opérateur, le commanditaire et le valideur.}
\usage{
mef_ajouter_intervenants(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des
différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter à la passerelle les intervenants sur les opérations
    Ces intervenants sont l'opérateur, le commanditaire et le valideur.
}
\examples{
\dontrun{
passerelle <- mef_creer_passerelle() \%>\%
mef_ajouter_intervenants()
}

}
