% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_ipr.R
\name{mef_ajouter_ipr}
\alias{mef_ajouter_ipr}
\title{Ajouter les notes IPR à une table passerelle}
\usage{
mef_ajouter_ipr(passerelle)
}
\arguments{
\item{passerelle}{Dataframe mettant en correspondance les identifiants des différentes tables.}
}
\value{
Dataframe avec la valeur de l'IPR, et la classe de qualité.
}
\description{
Ajouter les notes IPR à une table passerelle
}
\examples{
\dontrun{
mon_df_ipr <- mef_ajouter_ipr(passerelle = passerelle)
}

}
