% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_libelle.R
\name{mef_ajouter_libelle}
\alias{mef_ajouter_libelle}
\title{Rajouter le libellé du point (ou de la station quand il m'est pas dispobible au point) à
    un dataframe qui peut être une passerelle.}
\usage{
mef_ajouter_libelle(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle à laquelle ont été ajoutés les libellés des points de prélèvement.
}
\description{
Rajouter le libellé du point (ou de la station quand il m'est pas dispobible au point) à
    un dataframe qui peut être une passerelle.
}
\examples{
\dontrun{
passerelle <- mef_ajouter_libelle(passerelle = passerelle)
}
}
