% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_libelle_site.R
\name{mef_ajouter_libelle_site}
\alias{mef_ajouter_libelle_site}
\title{Rajouter le libellé du site}
\usage{
mef_ajouter_libelle_site(df, origine_libelle = "auto")
}
\arguments{
\item{df}{Dataframe qui, selon le choix de l'argument "origine_libelle", doit contenir les variables
"sta_id" et/ou "pop_id".}

\item{origine_libelle}{Caractère. Peut prendre les valeurs "point_wama", "station_sandre", ou "auto"
car la base Aspe contient différents champs indiquant la localisation. La valeur par défaut est "auto",
qui assure que le champ ne reste jamais vide (nécessaire pour certains traitements). Dans ce cas,
le champ retourné est nommé "pop_libelle" et sa valeur est par priorités décroissantes
"pop_libelle", "sta_libelle_sandre" et enfin "pop_id".}
}
\value{
Le dataframe auquel ont été ajoutés les libellés.
}
\description{
Rajoute à un dataframe le libellé du site, qui peut être le point ou la station.
}
\details{
Cette fonction nécessite que les dataframes de la base Aspe soient chargés dans l'environnement.
}
\examples{
\dontrun{
# chargement des packages
library(dplyr)
library(aspe)

# chargement des données (à adapter)
load(file = "raw_data/tables_sauf_mei_2023_04_07_09_39_32.RData")

# création du tableau
aspe_df <- mef_creer_passerelle() \%>\%
  mef_ajouter_libelle_site(origine_libelle = "station_sandre")
}
}
