% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_mei.R
\name{mef_ajouter_mei}
\alias{mef_ajouter_mei}
\title{Ajouter les mesures individuelles à la passerelle}
\usage{
mef_ajouter_mei(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
Elle doit contenir le champ "lop_id" qui sert à la jointure avec la table mesure_individuelle.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter les mesures individuelles à la passerelle
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_mei()
}
}
