% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_metriques.R
\name{mef_ajouter_metriques}
\alias{mef_ajouter_metriques}
\title{Ajouter à une passerelle les métriques IPR observées, théoriques et leur contribution
    à l'indice IPR agrégé.}
\usage{
mef_ajouter_metriques(passerelle)
}
\arguments{
\item{passerelle}{Le dataframe passerelle.}
}
\value{
Le dataframe passerelle complété par les métriques.
}
\description{
Ajouter à une passerelle les métriques IPR observées, théoriques et leur contribution
    à l'indice IPR agrégé.
}
\examples{
\dontrun{
passerelle_avec_metriques <- mef_ajouter_metriques(passerelle = ma_passerelle)
}
}
