% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_moyen_prospection.R
\name{mef_ajouter_moyen_prospection}
\alias{mef_ajouter_moyen_prospection}
\title{Ajouter à la passerelle le moyen de prospection}
\usage{
mef_ajouter_moyen_prospection(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ex : A pied / En bateau. Les modalités sont précisées dans la table ref_moyen_prospection.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_moyen_prospection()
}
}
