% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_objectif.R
\name{mef_ajouter_objectif}
\alias{mef_ajouter_objectif}
\title{Ajouter à la passerelle l'objectif de l'opération}
\usage{
mef_ajouter_objectif(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Par objectif, on entent la contribution à un réseau (DCE, RCS, RRP, etc.), le suivi de poopulations,
    des études, du suivi de restauration, etc. Les modalités sont précisées dans la table ref_objectif.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_objectif()
}
}
