% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_ope_date.R
\name{mef_ajouter_ope_date}
\alias{mef_ajouter_ope_date}
\title{Rajouter à la passerelle la date et l'année de chaque opération.}
\usage{
mef_ajouter_ope_date(df)
}
\arguments{
\item{df}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
df doit contenir une variable "ope_id" (identifiant de l'opération de pêche).}
}
\value{
La passerelle à laquelle ont été ajoutés la date et l'année.
}
\description{
Rajouter à la passerelle la date et l'année de chaque opération.
}
\examples{
\dontrun{
passerelle <- mef_ajouter_ope_date(df = passerelle)
}
}
