% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_ope_env.R
\name{mef_ajouter_ope_env}
\alias{mef_ajouter_ope_env}
\title{Rajouter les variables environnementales pour chaque opération à la passerelle (quand disponible)}
\usage{
mef_ajouter_ope_env(df)
}
\arguments{
\item{df}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle à laquelle ont été ajoutés les variables environnementales.
}
\description{
Ces variables sont en premier lieu renseignées depuis la table "operation_ipr", faute de quoi elles
sont complétées depuis la table "point_prelevement".
}
\examples{
\dontrun{
passerelle <- mef_ajouter_ope_env(df = passerelle)
}
}
