% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_operateur.R
\name{mef_ajouter_operateur}
\alias{mef_ajouter_operateur}
\title{Ajouter l'opérateur de pêche à la passerelle}
\usage{
mef_ajouter_operateur(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter l'opérateur de pêche à la passerelle
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_operateur()
}
}
