% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_passage.R
\name{mef_ajouter_passage}
\alias{mef_ajouter_passage}
\title{Ajouter le numéro de passage associé à chaque prélèvement élémentaire de la passerelle.}
\usage{
mef_ajouter_passage(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter le numéro de passage associé à chaque prélèvement élémentaire de la passerelle.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_passage()
}
}
