% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_qualification.R
\name{mef_ajouter_qualification}
\alias{mef_ajouter_qualification}
\title{Ajouter à la passerelle la qualification de a donnée
    ex : Correcte / Incorrecte / Incertaine.}
\usage{
mef_ajouter_qualification(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter à la passerelle la qualification de a donnée
    ex : Correcte / Incorrecte / Incertaine.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_qualification()
}
}
