% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_stats_taille.R
\name{mef_ajouter_stats_taille}
\alias{mef_ajouter_stats_taille}
\title{Ajouter les tailles (longueurs) mini, maxi, médiane et moyenne pour chaque espèce à
chaque opération}
\usage{
mef_ajouter_stats_taille(df)
}
\arguments{
\item{df}{Dataframe avec des variables espèce (esp_code_alternatif) et opération (ope_id).}
}
\value{
La passerelle à laquelle ont été ajoutés les statistiques de tailles.
}
\description{
La fonction fait appel à de nombreux dataframes de la base Aspe qui doivent impérativement avoir
été chargés auparavant (mesure_individuelle, lots, operation, etc.). Elle n'opère pas de distinction
entre longueur totale et longueur fourche.
}
\examples{
\dontrun{
passerelle <- mef_ajouter_stats_taille(passerelle = passerelle)
}
}
