% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_surf_calc.R
\name{mef_ajouter_surf_calc}
\alias{mef_ajouter_surf_calc}
\title{Rajouter la superficie échantillonnée calculée pour chaque opération à la passerelle}
\usage{
mef_ajouter_surf_calc(df)
}
\arguments{
\item{df}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
df doit contenir une variable "ope_id" (identifiant de l'opération de pêche).}
}
\value{
La passerelle à laquelle a été ajoutée la superficie échantillonnée qui est une valeur calculée
stockée dans le champ ope_surf_calculee de la table operation.
}
\description{
Rajouter la superficie échantillonnée calculée pour chaque opération à la passerelle
}
\examples{
\dontrun{
passerelle <- mef_ajouter_surf_calc(df = passerelle)
}
}
