% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_type_longueur.R
\name{mef_ajouter_type_longueur}
\alias{mef_ajouter_type_longueur}
\title{Ajouter le type de longueur à la passerelle}
\usage{
mef_ajouter_type_longueur(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
ex : totale, fourche.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_type_longueur()
}
}
