% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_type_materiel.R
\name{mef_ajouter_type_materiel}
\alias{mef_ajouter_type_materiel}
\title{Ajouter à la passerelle le type de matériel employé
    ex : Héron}
\usage{
mef_ajouter_type_materiel(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter à la passerelle le type de matériel employé
    ex : Héron
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_type_materiel()
}
}
