% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_type_protocole.R
\name{mef_ajouter_type_protocole}
\alias{mef_ajouter_type_protocole}
\title{Ajouter le type de protocole à la passerelle
    ex : Indice Abondance Saumon.}
\usage{
mef_ajouter_type_protocole(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter le type de protocole à la passerelle
    ex : Indice Abondance Saumon.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_type_protocole()
}
}
