% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_utilisateurs.R
\name{mef_ajouter_utilisateurs}
\alias{mef_ajouter_utilisateurs}
\title{Ajouter à la passerelle les utilisateurs sur les opérations
    Ces utilisateurs sont le créateur et la personne qui a effectué la dernière modification.}
\usage{
mef_ajouter_utilisateurs(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des
différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter à la passerelle les utilisateurs sur les opérations
    Ces utilisateurs sont le créateur et la personne qui a effectué la dernière modification.
}
\examples{
\dontrun{
passerelle <- mef_creer_passerelle() \%>\%
mef_ajouter_utilisateurs()
}

}
