% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_colo_ext_pop.R
\name{mef_colo_ext_pop}
\alias{mef_colo_ext_pop}
\title{Mettre en forme le tableau des captures sur un point}
\usage{
mef_colo_ext_pop(df, id_point)
}
\arguments{
\item{df}{Dataframe contenant les captures et contenant les champs pop_id, pop_libelle,
annee, esp_code_alternatif, effectif.}

\item{id_point}{Caractère. Identifiant du point au sens du champ pop_id de la base Aspe.}
}
\value{
Un dataframe pour un point de prélèvement qui est mis en forme pour produire un
    graphique avec la fonction gg_colo_ext_pop().
}
\description{
L'objectif est de visualiser la dynamique du peuplement. Permet d'identifier les taxons
    qui apparaissent ou disparaissent ainsi que les abondances.
}
\examples{
\dontrun{
colo_ext_mon_point <- captures \%>\%
    mef_colo_ext_pop(id_point = "35803"))
}
}
