% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_colo_ext_sta.R
\name{mef_colo_ext_sta}
\alias{mef_colo_ext_sta}
\title{Mettre en forme le tableau des captures sur un point en vue de visualiser la dynamique
    du peuplement. Permet d'identifier les taxons qui apparaissent ou disparaissent
    ainsi que les abondances.}
\usage{
mef_colo_ext_sta(df, id_station)
}
\arguments{
\item{df}{Dataframe contenant les captures et contenant les champs sta_id, pop_libelle,
annee, esp_code_alternatif, effectif.}

\item{id_station}{Caractère. Identifiant du point au sens du champ pop_id de la base Aspe.}
}
\value{
Un dataframe pour un point de prélèvement qui est mis en forme pour produire un
    graphique avec la fonction gg_colo_ext_sta().
}
\description{
Mettre en forme le tableau des captures sur un point en vue de visualiser la dynamique
    du peuplement. Permet d'identifier les taxons qui apparaissent ou disparaissent
    ainsi que les abondances.
}
\examples{
\dontrun{
colo_ext_ma_station <- captures \%>\%
    mef_colo_ext_sta(id_station = "22980"))
}
}
