% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_compter_pres_abs_env.R
\name{mef_compter_pres_abs_env}
\alias{mef_compter_pres_abs_env}
\title{Dénombrer le nb d'opérations avec chaque paramètre environnemental renseigné pour chacune
des espèces}
\usage{
mef_compter_pres_abs_env(df)
}
\arguments{
\item{df}{Dataframe avec les variables valeur_parametre, esp_code_alternatif, presence, parametre, annee, n.}
}
\value{
Un dataframe de 4 colonnes correspondant à l'espèce, au paramètre et aux nombres de présence et d'absence.
}
\description{
Dénombrer le nb d'opérations avec chaque paramètre environnemental renseigné pour chacune
des espèces
}
\examples{
\dontrun{
df_nb_pres_abs_par_parametre_env <- mef_compter_pres_abs_env(df = df_captures_et_env)
}
}
