% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_id.R
\name{mef_filtrer_id}
\alias{mef_filtrer_id}
\title{Filtrer les prélèvements}
\usage{
mef_filtrer_id(
  df,
  tb,
  var_id,
  var_sta = NULL,
  var_pro = NULL,
  min_obs,
  max_na_cons = NULL,
  max_pro = NULL,
  max_chg = NULL
)
}
\arguments{
\item{df}{data frame}

\item{tb}{data frame}

\item{var_id}{variable(s) identifiant les prélèvements}

\item{var_sta}{variable identifiant les points de prélèvements (optionnelle)}

\item{var_pro}{variable identifiant le protocole de pêche (optionnelle)}

\item{min_obs}{numerique définissant le nombre minimal d'observations par prélèvement}

\item{max_na_cons}{numerique définissant le nombre maximal de valeurs manquantes consécutives par prélèvement}

\item{max_pro}{numerique définissant le nombre maximal de protocoles de pêche par prélèvement}

\item{max_chg}{numerique définissant le nombre maximal de changement de protocole de pêche par prélèvement}
}
\value{
df
}
\description{
Filtrer les prélèvements
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
mef_ajouter_ope_date() \%>\%
mef_ajouter_ope_saison() \%>\%
mef_ajouter_type_protocole()
tb <- def_attribut_id (df,
var_id = c(pop_id,saison),
var_tmp = annee,
var_pro = pro_libelle)
df_id <- mef_filtrer_id(df,
tb,
var_id = c(pop_id, saison),
var_sta = pop_id,
var_pro = pro_libelle,
min_obs = 10,
max_na_cons = 3,
max_pro = 1)
df_id <- mef_filtrer_id(df,
tb,
var_id = pop_id,
min_obs = 10,
max_na_cons = 3)
}
}
