% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_ope_id.R
\name{mef_filtrer_ope_id}
\alias{mef_filtrer_ope_id}
\title{Filtrer les opérations de pêche}
\usage{
mef_filtrer_ope_id(
  df,
  var_id,
  var_tmp,
  var_surf = NULL,
  var_pro = NULL,
  var_pds = NULL,
  var_mei = NULL,
  var_pas = NULL,
  var_date = NULL,
  default = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{var_id}{variable(s) identifiant les prélèvements}

\item{var_tmp}{variable identifiant la date de prélèvement}

\item{var_surf}{variable identifiant la surface de pêche (optionnelle)}

\item{var_pro}{variable identifiant le protocole de pêche (optionnelle)}

\item{var_pds}{variable identifiant la présence d'information sur les poids (optionnelle)}

\item{var_mei}{variable identifiant la présence d'information sur les mesures individuelles (optionnelle)}

\item{var_pas}{variable identifiant le numéro du passage de pêche (optionnelle)}

\item{var_date}{variable identifiant une date additionnelle de pêche (optionnelle)}

\item{default}{logique}
}
\value{
df
}
\description{
Filtrer les opérations de pêche
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
mef_ajouter_ope_date() \%>\%
mef_ajouter_surf_calc() \%>\%
mef_ajouter_type_protocole() \%>\%
mef_filtrer_ope_id(var_id = pop_id,
var_tmp = annee,
var_surf = ope_surface_calculee,
var_pro = pro_libelle,
default=TRUE)
df <- mef_creer_passerelle() \%>\%
mef_ajouter_ope_date() \%>\%
mef_ajouter_surf_calc() \%>\%
mef_ajouter_type_protocole() \%>\%
mef_ajouter_poids() \%>\%
mef_filtrer_ope_id(var_id = pop_id,
var_tmp = annee,
var_surf = ope_surface_calculee,
var_pds = lop_poids)
df <- df \%>\%
mef_filtrer_ope_id(var_id = pop_id,
var_tmp = annee,
var_pro = pro_libelle,
default=TRUE)
}
}
