% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_sta_annee.R
\name{mef_filtrer_sta_annee}
\alias{mef_filtrer_sta_annee}
\title{Filtrer un jeu de données sur les années et les identifiants des stations.}
\usage{
mef_filtrer_sta_annee(
  ipr_df,
  stations_id = NA,
  premiere_annee = NA,
  derniere_annee = NA
)
}
\arguments{
\item{stations_id}{Vecteur caractère contenant les identifiants (sta_id) des stations à
représenter sur le graphique.}

\item{premiere_annee}{Numérique. Année de début de la période à représenter (incluse).
Par défaut, c'est la première année de donnée sur l'ensemble des stations.}

\item{derniere_annee}{Numérique. Année de fin de la période à représenter (incluse).
Par défaut, c'est la dernière année de donnée sur l'ensemble des stations.}

\item{df}{Dataframe contenant les données, issu de la fonction extraire_ipr() ou au moins
avec des colonnes "annee" et "sta_id".}
}
\value{
Le dataframe filtré.
}
\description{
Filtrer un jeu de données sur les années et les identifiants des stations.
}
\examples{
\dontrun{
ipr_depuis_2010 <- mef_filtrer_sta_annee(df = data_22,
premiere_annee = 2010)
}
}
