% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_validation.R
\name{mef_filtrer_validation}
\alias{mef_filtrer_validation}
\title{Filtrer un dataframe passerelle sur le niveau de validation des opérations.}
\usage{
mef_filtrer_validation(df, validation = 3)
}
\arguments{
\item{df}{Dataframe passerelle qui doit contenir le champ "ope_id".}

\item{validation}{Vecteur entier indiquant les modalités à sélectionner, selon le codage
du champ eta_id de la table ref_etat_avancement.
1 = "En cours de saisie"
2 = "Validé niveau 1"
3 = "Validé niveau 2" (Valeur par défaut).}
}
\value{
Dataframe filtré, c'est-à-dire expurgé de certaines opérations.
}
\description{
Filtrer un dataframe passerelle sur le niveau de validation des opérations.
}
\examples{
\dontrun{
# Opérations qui ne sont plus en cours de saisie
data_22_filtre <- mef_filtrer_validation(df = data_22, validation = 2:3)
}
}
