% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_pivoter_var_env.R
\name{mef_pivoter_var_env}
\alias{mef_pivoter_var_env}
\title{Pivoter le dataframe pour regrouper les variables environnementales sur 2 colonnes}
\usage{
mef_pivoter_var_env(df)
}
\arguments{
\item{df}{Dataframe avec les variables distance_mer, temp_janvier, etc.}
}
\value{
Le dataframe pivoté, en forme pour les graphiques avec ggplot2.
}
\description{
Pivoter le dataframe pour regrouper les variables environnementales sur 2 colonnes
}
