% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_recoder_esp_code_alt.R
\name{mef_recoder_esp_code_alt}
\alias{mef_recoder_esp_code_alt}
\title{Recodage des codes taxons 'esp_code_alternatif'}
\usage{
mef_recoder_esp_code_alt(df, type_recodage = "especes")
}
\arguments{
\item{df}{Dataframe contenant la variable esp_code_alternatif'.}

\item{type_recodage}{Caractère, soit "especes" (par défaut), soit "ipr".
Dans le premier cas, les codes taxons de sous-espèces ou écotype sont recodés à l'espèce.
Dans le second cas, les codes taxons sont recodés selon la taxonomie de l'IPR, ici
conformément au script 'IPR_v1.0.3_calc_consult.R'
téléchargé depuis [https://seee.eaufrance.fr/](https://seee.eaufrance.fr/).}
}
\value{
Le dataframe mis à jour.
}
\description{
Recodage des codes taxons 'esp_code_alternatif'
}
\examples{
\dontrun{
# Calcul des captures par espèces pour chaque opération de pêche
library(dplyr)
library(aspe)

load(file = "raw_data/tables_sauf_mei_2023_04_07_09_39_32.RData")

captures <- mef_creer_passerelle() \%>\%
 mef_ajouter_lots() \%>\%
 mef_ajouter_esp() \%>\%
 mef_recoder_esp_code_alt() \%>\%
 group_by(ope_id, esp_code_alternatif) \%>\%
 summarise(effectif = sum(lop_effectif)) \%>\%
 ungroup()
}
}
