% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_recoder_filtrer_esp.R
\name{mef_recoder_filtrer_esp}
\alias{mef_recoder_filtrer_esp}
\title{Recodage des codes espèces et exclusion de taxons}
\usage{
mef_recoder_filtrer_esp(df, esp_a_supprimer = NA)
}
\arguments{
\item{df}{Dataframe standardisé contenant les données.}

\item{esp_a_supprimer}{Vecteur texte contenant les codes à trois lettres des taxons à
supprimer. Il contient nécessairement un champ "esp_code_alternatif".}
}
\value{
Le dataframe mis à jour.
}
\description{
Par exemple les carpes cuir, miroir, etc. sont regroupées sous un unique code CCX.
Idem pour les vandoises en VAX.
}
\examples{
\dontrun{
df_propre <- df_brut \%>\%
mef_recoder_filtrer_esp(esp_a_supprimer = c("OCV", "ASA"))
}
}
