% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_select_dept.R
\name{mef_select_dept}
\alias{mef_select_dept}
\title{Filtrer la passerelle sur un ou des département(s).}
\usage{
mef_select_dept(passerelle, dept)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}

\item{dept}{Vecteur numérique indiquant le ou les département(s) à conserver.}
}
\value{
La passerelle filtrée sur le ou les département(s).
}
\description{
Filtrer la passerelle sur un ou des département(s).
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_select_dept(dept = c(22, 29, 35, 56))
}
}
