% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_select_mop.R
\name{mef_select_mop}
\alias{mef_select_mop}
\title{Filtrer la passerelle sur le moyen de prospection.}
\usage{
mef_select_mop(passerelle, moyen_prospection)
}
\arguments{
\item{passerelle}{Dataframe "passerelle mettant en correspondance les identifiants des différentes tables.}

\item{moyen_prospection}{Vecteur numérique indiquant le ou les moyens de prospection
à conserver. A choisir parmi 1 = "A pied", 2 = "En bateau", 3 = "Mixte" et 4 = "Non renseigné".}
}
\value{
La passerelle filtrée sur le ou les moyens de prospection.
}
\description{
Filtrer la passerelle sur le moyen de prospection.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_select_mop(moyen_prospection = c(1, 2, 3))
}
}
