% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_select_obj.R
\name{mef_select_obj}
\alias{mef_select_obj}
\title{Filtrer la passerelle sur le ou les objectif(s) poursuivis par l'opération. On appelle objectif
    par exemple l'appartenance à un réseau.}
\usage{
mef_select_obj(passerelle, objectif)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}

\item{objectif}{Vecteur numérique indiquant le ou les objectif(s) à conserver.
A choisir parmi :
1 = "Sauvetage - Transfert",
2 = "DCE – Référence",
3 = "RCS – Réseau de Contrôle de Surveillance",
4 = "RRP – Réseau de Référence Pérenne",
5 = "CO – Contrôle opérationnel",
6 = "RHP – Réseau Hydrobiologique Piscicole",
etc.
La correspondance entre les codes et les libellés peut être retrouvée dans
la table ref_objectif.}
}
\value{
La passerelle filtrée sur le ou les objectif(s).
}
\description{
Filtrer la passerelle sur le ou les objectif(s) poursuivis par l'opération. On appelle objectif
    par exemple l'appartenance à un réseau.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_select_obj(objectif = 2:6)
}
}
