% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_select_proto.R
\name{mef_select_proto}
\alias{mef_select_proto}
\title{Filtrer la passerelle sur le protocole employé.}
\usage{
mef_select_proto(passerelle, protocole)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}

\item{protocole}{Vecteur numérique indiquant le ou les protocole(s) à conserver.
A choisir parmi :
1 = "Pêche complète à un ou plusieurs passages",
2 = "Pêche partielle par points (grand milieu)",
3 = "Pêche par ambiances",
4 = "Pêche partielle sur berge",
5 = "Indice Abondance Saumon",
6 = "Vigitruite",
7 = "Indice Abondance Anguille"
La correspondance entre les codes et les libellés peut être retrouvée dans
la table ref_protocole.}
}
\value{
La passerelle filtrée sur le ou les protocole(s).
}
\description{
Filtrer la passerelle sur le protocole employé.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_select_proto(protocole = c(1, 3))
}
}
