% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ssech_abs.R
\name{mef_ssech_abs}
\alias{mef_ssech_abs}
\title{Sous-échantillonner parmi les absences}
\usage{
mef_ssech_abs(df, seed = 123)
}
\arguments{
\item{df}{Dataframe avec les variables valeur_parametre, esp_code_alternatif, presence, parametre, n.}

\item{seed}{Numérique. La racine du générateur de nombres aléatoires pour le tirage.}
}
\value{
Un dataframe issu de df mais avec moins de lignes.
}
\description{
La comparaison entre les sites de présence et d'absence d'une espèce n'a de sens quer si leur nombre
est comparable. IL ne sert à rien d'avoir l'altitude de 100 000 sites où le saumon est absent
si l'on n'a que 1000 sites où il est présent. On allège donc le dataframe en tirant aléatoirement
autant d'opérations d'absence qu'il y a de présence.
}
\examples{
\dontrun{
df_allege <- mef_ssech_abs(df = df_pres_abs_env)
}
}
