% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_charger_donnees_test.R
\name{misc_charger_donnees_test}
\alias{misc_charger_donnees_test}
\title{Charger un extrait de la base Aspe qui permet d'effectuer des tests}
\usage{
misc_charger_donnees_test()
}
\value{
Les dataframes dans l'environnement de travail.
}
\description{
Ces données servent aussi aux vignettes.
}
\examples{
\dontrun{
# chargement des données
misc_charger_donnees_test()

# vérification que les dataframes sont bien accessibles
ls()

# assemblage du jeu de données
passerelle <- aspe::mef_creer_passerelle()
}
}
