% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_derniere_date.R
\name{misc_derniere_date}
\alias{misc_derniere_date}
\title{Obtenir la date maxi saisie dans la base Aspe}
\usage{
misc_derniere_date()
}
\value{
La date maximale trouvée
}
\description{
Cette fonction extrait l'ensemble des champs contenant la chaîne de caractère "date"
dans l'ensemble des dataframes, en prend le max et omettant les dates ultérieures
à la date système.
}
\examples{
\dontrun{
date_max <- misc_derniere_date()
}
}
