% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_nom_dernier_fichier.R
\name{misc_nom_dernier_fichier}
\alias{misc_nom_dernier_fichier}
\title{Obtenir le nom du plus récent des fichiers}
\usage{
misc_nom_dernier_fichier(repertoire = NULL, pattern, recursive = FALSE)
}
\arguments{
\item{repertoire}{Chaîne de caractère. Chemin vers le répertoire à explorer.
Par défaut c'est le répertoire de travail.}

\item{pattern}{Chaîne de caractère (exprtession régulière) à retrouver dans le nom du fichier.}

\item{recursive}{Booléen (TRUE ou FALSE). Indique si la recherche inclut les sous-répertoire.
La valeur par défaut est FALSE donc les sous-répertoires sont exclus de la recherche.}
}
\value{
Le nom du plus récent des fichiers respectant la condition indiquée au pattern.
}
\description{
Recherche dans un répertoire, parmi les fichiers dont le nom correspond au pattern,
celui qui est le dernier créé.
}
\examples{
\dontrun{
mon_fichier <- misc_nom_dernier_fichier(repertoire = "raw_data", pattern = "RData$")
}
}
