% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passer_ipr_large.R
\name{passer_ipr_large}
\alias{passer_ipr_large}
\title{Convertir une table des IPR du format long au format large, c'est-à-dire avec une colonne
    par année.
S'il y a plisueurs valeurs de l'IPR pour un même point une année donnée, la valeur retournée
    sera la moyenne}
\usage{
passer_ipr_large(ipr_df)
}
\arguments{
\item{ipr_df}{Le dataframe au format long contenant les IPR, issi de la fonction extraire-ipr()}
}
\value{
Un dataframe contenant les mêmes données, mais agencé avec une colonne par année.
}
\description{
Convertir une table des IPR du format long au format large, c'est-à-dire avec une colonne
    par année.
S'il y a plisueurs valeurs de l'IPR pour un même point une année donnée, la valeur retournée
    sera la moyenne
}
\examples{
\dontrun{
ipr_large_df <- passer_ipr_large(ipr_df = mon_df_ipr)
}
}
