% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtp_estimer_poids.R
\name{qtp_estimer_poids}
\alias{qtp_estimer_poids}
\title{Calculer les poids manquants à partir des longueurs}
\usage{
qtp_estimer_poids(df_biometrie, espece, type_longueur = 2, df_taille_poids)
}
\arguments{
\item{df_biometrie}{Dataframe contenant les données de biométrie avec en particulier
les champs esp_code_alternatif, tlo_id, mon_lop_poids_est et mei_taille.}

\item{espece}{Caractère. Code espèce en trois lettres.}

\item{type_longueur}{Numérique.
1 = Longueur fourche
2 = Longueur totale}

\item{df_taille_poids}{Dataframe contenant les paramètres de la relation taille-poids.
Il doit avoir les champs code_espece a et b.}
}
\value{
Un dataframe complété dans le champ mon_lop_poids_est pour l'espèce considérée
    avec le type de longueur considéré.
}
\description{
Pour une espèce et un type de longueur.
}
\examples{
\dontrun{
essai <- qtp_calculer_poids(df = lop_poids_est, espece = "GOU",
type_longueur = 2, df_taille_poids = tp)
}
}
