% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtp_histo.R
\name{qtp_histo}
\alias{qtp_histo}
\title{Produire l'histogramme des longueurs ou des masses d'une espèce.}
\usage{
qtp_histo(df, code_espece, variable = "mei_taille", x_min = NA, x_max = NA)
}
\arguments{
\item{df}{Dataframe contenant les données}

\item{code_espece}{Texte. Code à trois lettres de l'espèce.}

\item{variable}{Texte. La variable à représenter, soit "mei_taille", soit "mei_poids".}

\item{x_min}{Minimum de l'axe des abscisses.}

\item{x_max}{Maximum de l'axe des abscisses.}
}
\value{
Un graphique ggplot.
}
\description{
Produire l'histogramme des longueurs ou des masses d'une espèce.
}
\examples{
\dontrun{
qtp_histo (df = tp_data,
code_espece = "ROT",
variable = "mei_taille",
x_max = 500)
}

}
