% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtp_nettoyer_fb_lieu.R
\name{qtp_nettoyer_fb_lieu}
\alias{qtp_nettoyer_fb_lieu}
\title{Nettoyer les intitulés de lieux issus de Fishbase}
\usage{
qtp_nettoyer_fb_lieu(df)
}
\arguments{
\item{df}{Le dataframe contenant les paramètres taille-poids et une variable "lieu" à
nettoyer. Typiquement importé avec la fonction rfishbase::length_weight() puis
dont la variable "Locality" a été renommée en "lieu".}
}
\value{
Le dataframe avec la colonne "lieu" propre.
}
\description{
Nettoyer les intitulés de lieux issus de Fishbase
}
\examples{
\dontrun{
tp_fb_propre <- qtp_nettoyer_fb_lieu(df = tp_fb_brut)
}
}
