% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtp_resume_donnees.R
\name{qtp_resume_donnees}
\alias{qtp_resume_donnees}
\title{Résumer le contenu du tableau de données avant de construire la relation taille - poids.}
\usage{
qtp_resume_donnees(df, seuil_poids_absolu)
}
\arguments{
\item{df}{Dataframe contenant les données.}

\item{seuil_poids_absolu}{Numérique. Poids minimum pour qu'un individu ne soit pas exclus,
en grammes.}
}
\value{
Un dataframe de résumé.
}
\description{
Résumer le contenu du tableau de données avant de construire la relation taille - poids.
}
\examples{
\dontrun{
qtp_resume_donnees (df = tp_data,
seuil_poids_absolu = 10)
}
}
