% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtp_seuils.R
\name{qtp_seuils}
\alias{qtp_seuils}
\title{Obtenir les seuils mini et maxi des variables biométriques selon une condition de
    probabilité.}
\usage{
qtp_seuils(df, code_espece, variable = "mei_taille", seuil_densite = 0.01)
}
\arguments{
\item{df}{Dataframe contenant les données}

\item{code_espece}{Texte. Code à trois lettres de l'espèce.}

\item{variable}{Texte. La variable à représenter, soit "mei_taille", soit "mei_poids".}

\item{seuil_densite}{Numérique. Seuil probabilité associé aux intervalles de taille.}
}
\value{
Un vecteur contenant les seuils mini et maxi pour la variable considérée.
}
\description{
Obtenir les seuils mini et maxi des variables biométriques selon une condition de
    probabilité.
}
\examples{
\dontrun{
qtp_seuils (df = data_esp,
            code_espece = "TRF",
            variable = "mei_taille",
            seuil_densite = 0.001)
}
}
