% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renommer_pour_macro.R
\name{renommer_pour_macro}
\alias{renommer_pour_macro}
\title{Sélectionner et nommer les colonnes au format requis en entrée de la macro
    Excel de calcul de l'indice.}
\usage{
renommer_pour_macro(data, fichier_macro)
}
\arguments{
\item{data}{Le dataframe contenant les données, issu de la fonction formater_pour_macro().}

\item{fichier_macro}{Caractère. Chemin vers le fichier Excel de la macro.}
}
\value{
Le dataframe
}
\description{
Sélectionner et nommer les colonnes au format requis en entrée de la macro
    Excel de calcul de l'indice.
}
\examples{
\dontrun{
data_macro <- formater_pour_macro(data = data_macro, fichier_macro = "raw_data/CalculIPRv1.3.xls")
}
}
