% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scinder_dump.R
\name{scinder_dump}
\alias{scinder_dump}
\title{Scinder le dump en deux parties pour les traiter séparément et éviter de
    sature la mémoire vive}
\usage{
scinder_dump(fichier_dump, repertoire_sortie = NA)
}
\arguments{
\item{fichier_dump}{Le chemin vers le fichier d'extension .sql (décompressé) ou
.sql.gz (archive).}

\item{repertoire_sortie}{Le chemin vers le répertoire de sortie (qui doit avoir
été créé auparavant). Par défaut c'est celui qui contient le dump.}
}
\value{
Deux fichiers au format .RData contenant l'un les lignes de la table
    mesure_individuelle et l'autre tout le reste du dump
}
\description{
Scinder le dump en deux parties pour les traiter séparément et éviter de
    sature la mémoire vive
}
\examples{
\dontrun{
scinder_dump(fichier_dump = "../raw_data/dump.sql")
}

}
