% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-connexion-idg.R
\name{start_connexion_idg}
\alias{start_connexion_idg}
\title{Démarre une connexion à ASPE sur la BDD IDG (OFB)}
\usage{
start_connexion_idg()
}
\value{
Connexion. Renvoie une erreur si connexion impossible.
}
\description{
Fonction pour se connecter à la BDD ASPE en utilisant les paramètres de connexion
stockés dans l'environnement R (fichier \code{.Renviron}).
}
\examples{
\dontrun{
bdd_aspe <- start_connexion_idg()

start_connexion_idg(conn = bdd_aspe)
}

}
