% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_poissons_data.R
\docType{data}
\name{traits_poissons}
\alias{traits_poissons}
\title{Données sommaires sur les espèces de poissons}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\usage{
data(traits_poissons)
}
\description{
Tableau d'une centaine de lignes correspondant aux espèces de poissons
référencées par leur code à trois lettres et 5 colonnes précisant le
statut de ces espèces ainsi que leur principaux habitats.
}
\examples{
\donttest{
data(traits_poissons)
traits_poissons}
}
\keyword{datasets}
