% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_attribut_id.R
\name{def_attribut_id}
\alias{def_attribut_id}
\title{Définir les attributs des prélèvements}
\usage{
def_attribut_id(df, var_id, var_tmp, var_pro = NULL)
}
\arguments{
\item{df}{dataframe}

\item{var_id}{variable(s) identifiant les prélèvements}

\item{var_tmp}{variable identifiant la date de prélèvement}

\item{var_pro}{variable identifiant le protocole de pêche (optionnelle)}
}
\value{
tb data frame
}
\description{
Définir les attributs des prélèvements
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
mef_ajouter_ope_date() \%>\%
mef_ajouter_type_protocole()
def_id <- def_attribut_id (var_id = pop_id,
var_tmp = annee,
var_pro = pro_libelle)
}
}
