% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl_trouver_table.R
\name{expl_trouver_table}
\alias{expl_trouver_table}
\title{Rechercher les dataframes dont le nom contient un pattern}
\usage{
expl_trouver_table(nom_contient)
}
\arguments{
\item{nom_contient}{Chaine de caractères recherchés parmi les noms des tables.}
}
\value{
Un vecteur caractères contenant les noms des tables.
}
\description{
Cette fonction est utile pour explorer la base quand on a une idée de ce que son nom
peut contenir. Par exemple pour les tables liées au point on peut chercher, parmi les noms
des dataframes chargés, ceux contenant la chaîne de caractères "point".
}
\examples{
\dontrun{
expl_trouver_table(nom_contient = "point")
}
}
