% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tables_csv.R
\name{export_tables_csv}
\alias{export_tables_csv}
\title{Exporter tous les dataframes de l'environnement au format .csv}
\usage{
export_tables_csv(repertoire = NA)
}
\arguments{
\item{repertoire}{Caractères. Chemin vers le répertoire où les fichiers seront écrits.}
}
\value{
Les fichiers sont sauvegardés dans le répertoire choisi. Si aucun chemin n'est indiqué,
    ça sera dans un sous-répertoire "processed_data/csv" du répertoire de travail qui est créé au besoin.
}
\description{
Exporter tous les dataframes de l'environnement au format .csv
}
\examples{
\dontrun{
export_tables_csv()
}
}
