% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_dyn_esp_pop.R
\name{gg_dyn_esp_pop}
\alias{gg_dyn_esp_pop}
\title{Produire un graphique de synthèse de la dynamique d'une population}
\usage{
gg_dyn_esp_pop(df, point, espece)
}
\arguments{
\item{df}{Dataframe contenant les données, avec les variables "pop_id", "pop_libelle",
esp_nom_commun", "mei_taille" et "annee".}

\item{point}{Caractère. Nom du point dans le champ "pop_libelle".}

\item{espece}{Caractère. Nom de l'espèce dans le champ "esp_nom_commun".}
}
\value{
Le graphique ggplot2 avec les distributions en taille par année et l'évolution
    temporelle de la longueur médiane et des effectifs capturés.
}
\description{
Représentation pour une espèce sur un point de prélèvement.
}
\examples{
\dontrun{
gg_dyn_esp_pop(df = data,
               point = "DROME A CHABRILLAN ",
               espece = "Barbeau fluviatile")
}
}
