% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_env_fn_annee.R
\name{gg_env_fn_annee}
\alias{gg_env_fn_annee}
\title{Produire le graphique de l'évolution temporelle de facteurs environnementaux moyens comparés
    entre les sites de présence et d'absence des espèces.}
\usage{
gg_env_fn_annee(
  df,
  parametres = NA,
  log_y = c("surface_bv", "distance_source", "pente", "largeur"),
  annee_debut = 1995,
  annee_fin = NA,
  y_libre = FALSE
)
}
\arguments{
\item{df}{Le dataframe contenant les données avec les variables presence, parametre,
valeur_parametre, annee}

\item{parametres}{Vecteur caractères. Les paramètre (ex : "altitude") représentés.}

\item{log_y}{Vecteur caractères. Les paramètre (ex : "altitude") représentés en échelle log.
Par défaut, sa valeur est c('surface_bv', 'distance_source', 'pente', 'largeur'), ce qui
correspond aux paramètres dont la distribution de rapproche d'une log-normale.}

\item{annee_debut}{Entier. Année mini à représenter.}

\item{annee_fin}{Entier. Année maxi à représenter.}

\item{y_libre}{TRUE/FALSE (valeur par défaut). Les axes des ordonnées peuvent (TRUE) différer
selon les espèces, ce qui permet un "zoom", ou bien être les mêmes pour toutes, ce qui permet
de les comparer les unes aux autres.}
}
\value{
Une liste dont les éléments sont des graphiques ggplot2 (un par paramètre).
}
\description{
Produire le graphique de l'évolution temporelle de facteurs environnementaux moyens comparés
    entre les sites de présence et d'absence des espèces.
}
\examples{
\dontrun{
mes_graphiques <- gg_env_fn_annee(df = mon_df, log_y = c('surface_bv', 'distance_source'))
mes_graphiques[[1]]
}
}
