% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_temp_metriq_grille.R
\name{gg_temp_metriq_grille}
\alias{gg_temp_metriq_grille}
\title{Graphique de la série chronologique des métriques IPR sur une station ou un groupe de stations}
\usage{
gg_temp_metriq_grille(
  df_metriques,
  var_id_sta,
  var_nom_metrique,
  var_valeur_metrique,
  station_sel = NULL,
  nb_colonnes = 7,
  max_axe_y = 10,
  id_sta_max_caract = 25,
  inv_y = TRUE,
  orientation = FALSE
)
}
\arguments{
\item{df_metriques}{Dataframe contenant les données. Il doit contenir une variable "annee" ainsi que
trois variables pour d'identifier les stations (ou points de prélèvement) et les métriques, ainsi que
pour donner la valeur des métriques.}

\item{var_id_sta}{Variable servant à identifier les stations ou points.
Cette variable donnera les étiquettes des lignes du graphique.}

\item{var_nom_metrique}{Variable contenant les noms des métriques (ex : dio, dti).}

\item{var_valeur_metrique}{Variable numérique contenant les valeurs des métriques.}

\item{station_sel}{Vecteur caractère indiquant les points ou stations à sélectionner.}

\item{nb_colonnes}{Entier. Nombre de colonnes du graphique. Par défaut nb_colonnes = 7 pour les 7 métriques IPR.
Dans le cas où une seule station est sélectionnée, et seulement dans ce cas, nb_colonnes peut être différent de 7.}

\item{max_axe_y}{Numérique. Limite supérieure de l'axe des ordonnées. Par défaut max_axe_y = 10.}

\item{id_sta_max_caract}{Entier. Nombre maximum de caractères dans l'identifiant de la station, au-delà duquel il sera
découpé pour tenir sur plusieurs lignes. Par défaut c'est 25 caractères.}

\item{inv_y}{Booléen. Indique l'axe des ordonnées pointe vers le bas (TRUE, par défaut) ou
vers le haut. NB pour l'IPR, plus l'indice est faible plus la qualité est élevée.
C'est l'inverse pour l'IPR+.}

\item{orientation}{Caractère. Par défaut les métriques sont organisées horizontalement (orientation = "h"). Pour permettre
d'organiser les métriques en 2 colonnes correspondant aux métriques de richesse et de densité, il faut
spécifier orientation = "v". Cet argument ne fonctionne que si une seule station est sélectionnée.}
}
\value{
Un graphique ggplot2.
}
\description{
Cette fonction renvoie un graphique ggplot2 en treillis avec une colonne par métrique IPR et une ligne par station.
}
\examples{
\dontrun{
# préparation des données
metriques <- mef_creer_passerelle() \%>\%
select(-lop_id, -pre_id) \%>\%
  distinct() \%>\%
  mef_ajouter_metriques() \%>\%
  mef_ajouter_libelle() \%>\%
  mef_ajouter_ope_date() \%>\%
  filter(!is.na(ner)) \%>\%
  select(-ends_with("observe"), -ends_with("theorique")) \%>\%
  pivot_longer(cols = ner:dti,
               names_to = "metrique",
               values_to = "valeur")

# affichage
gg_temp_metriq_grille(df_metriques = metriques,
                      station_sel = c("La Berre à Portel-des-Corbières", "LA BERENCE A GAVRAY"),
                      var_id_sta = pop_libelle,
                      var_nom_metrique = metrique,
                      var_valeur_metrique = valeur)

gg_temp_metriq_grille(df_metriques = metriques,
                      station_sel = c("La Berre à Portel-des-Corbières"),
                      var_id_sta = pop_libelle,
                      var_nom_metrique = metrique,
                      var_valeur_metrique = valeur,
                      nb_colonnes = 2,
                      orientation = "v"
)
}
}
