% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneiser_coordonnees.R
\name{homogeneiser_coordonnees}
\alias{homogeneiser_coordonnees}
\title{Homogénéiser les systèmes de coordonnées}
\usage{
homogeneiser_coordonnees(df, crs_sortie = 4326)
}
\arguments{
\item{df}{Le dataframe contenant des coordonnées (sta_coordonnees_x et sta_coordonnees_y)
ainsi qu'une colonne avec le système de coordonnées}

\item{crs_sortie}{Numérique. Le numéro EPSG du CRS de sortie. Les principaux sont :
Lambert II Etendu : crs = 27572
RGF93 / Lambert 93 : crs = 2154}
}
\value{
Le dataframe spatial (classe sf) dans le CRS choisi en sortie. Par défaut
    c'est en WGS84 (crs = 4326).
}
\description{
La base Aspe comprend deux systèmes de coordonnées : Lambert II étendu et
    Lambert 93. Il s'agit ici d'homogénéiser.
}
\examples{

\dontrun{
homogeneiser_coordonnees(df = station, crs_sortie = 4326)
}
}
