% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_corres_aspe_taxref.R
\name{imp_corres_aspe_taxref}
\alias{imp_corres_aspe_taxref}
\title{Importer une table de correspondance entre les codes espèces 3 lettres Aspe et leurs codes Taxref}
\usage{
imp_corres_aspe_taxref(
  url =
    "https://api.sandre.eaufrance.fr/referentiels/v1/apt.csv?compress=true&filter=<Filter><IS><Field>CodeAlternatifAp/OrgCdAlternatif</Field><Value>ASPE</Value></IS></Filter>"
)
}
\arguments{
\item{url}{Caractère. URL du jeu de données à télécharger.
Par défaut "https://api.sandre.eaufrance.fr/referentiels/v1/apt.csv?compress=true&
filter=\if{html}{\out{<Filter>}}\if{html}{\out{<IS>}}\if{html}{\out{<Field>}}CodeAlternatifAp/OrgCdAlternatif\if{html}{\out{</Field>}}\if{html}{\out{<Value>}}ASPE\if{html}{\out{</Value>}}\if{html}{\out{</IS>}}\if{html}{\out{</Filter>}}"}
}
\value{
Dataframe. Table de correspondence.
}
\description{
La donnée ne provient pas de la base ASPE, mais de l'API SANDRE. Les données importées ont été
filtrées afin de ne pas récupérer l'intégralité du référentiel "Appellation de taxons". Seuls
}
\examples{
\dontrun{
taxref <- imp_corres_aspe_taxref()
}
}
