% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_importer_aspe_idg.R
\name{imp_importer_aspe_idg}
\alias{imp_importer_aspe_idg}
\title{Importer les tables ASPE à partir de la BDD IDG (OFB)}
\usage{
imp_importer_aspe_idg(conn, ...)
}
\arguments{
\item{conn}{connexion à la BDD}

\item{...}{autre}
}
\value{
Les dataframes correspondant à chacune des tables de la base (sauf celles expurgées au préalable)
}
\description{
Importer les tables ASPE à partir de la BDD IDG (OFB)
}
\examples{
\dontrun{
bdd_aspe <- start_connexion_aspe_idg()

imp_importer_aspe_idg(conn = bdd_aspe)

stop_connexion_aspe_idg(conn = bdd_aspe)
}
}
