% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr_grapher_plusieurs_stations.R
\name{ipr_grapher_plusieurs_stations}
\alias{ipr_grapher_plusieurs_stations}
\title{Produire un graphique en treilis de l'évolution de l'IPR sur un jeu de stations.}
\usage{
ipr_grapher_plusieurs_stations(
  ipr_df,
  titre = NA,
  palette = NA,
  nb_colonnes = 3,
  max_axe_y = 40
)
}
\arguments{
\item{ipr_df}{Dataframe contenant les données, issu de la fonction extraire_ipr().}

\item{titre}{Caractère. Titre du graphique.}

\item{palette}{Vecteur nommé de couleurs, contenant 5 éléments pour chacune des
classes IPR. Par défaut, une palette est fournie.}

\item{nb_colonnes}{Nombre entier. Nombre de colonnes dans la sortie graphique.}

\item{max_axe_y}{Numérique. Valeur maximale de l'axe des ordonnées (IPR = 40 par défaut).
Ce paramètre peut être intéressant pour éviter que quelques valeurs très élevées
nuisent à la lisibilité du graphique.}
}
\value{
Le graphique issu de ggplot2.
}
\description{
Produire un graphique en treilis de l'évolution de l'IPR sur un jeu de stations.
}
\examples{
\dontrun{
ipr_grapher_plusieurs_stations(ipr_df = data_22,
nb_mini_annees = 3,
titre = "Morbihan")
}
}
