% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_esp_code_alternatif.R
\name{mef_ajouter_esp_code_alternatif}
\alias{mef_ajouter_esp_code_alternatif}
\title{Ajouter le code espèce à 3 lettres à la passerelle.}
\usage{
mef_ajouter_esp_code_alternatif(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
La passerelle doit comprendre un champ avec le code taxon esp_id (typiquement, on utilise donc
mef_ajouter_esp_code_alternatif() après mef_ajouter_lots())}
}
\value{
La passerelle complétée.
}
\description{
Il s'agit des codes "historiques" hérités du Conseil Supérieur de la Pêche.
}
\examples{
\dontrun{
passerelle <- mef_creer_passerelle() \%>\%
mef_ajouter_lots() \%>\%
mef_ajouter_esp_code_alternatif()
}
}
