% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_libelle.R
\name{mef_ajouter_libelle}
\alias{mef_ajouter_libelle}
\title{Rajouter le libellé du point}
\usage{
mef_ajouter_libelle(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle à laquelle ont été ajoutés les libellés des points de prélèvement.
}
\description{
Rajoute à un dataframe qui peut être une passerelle le libellé du point
    (ou de la station quand il m'est pas dispobible au point). Si ni l'un
    ni l'autre ne sont renseignés, c'est le pop_id pour éviter les valeurs
    manquantes bloquantes pour certains traitements.
}
\examples{
\dontrun{
passerelle <- mef_ajouter_libelle(passerelle = passerelle)
}
}
