% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_lots.R
\name{mef_ajouter_lots}
\alias{mef_ajouter_lots}
\title{Ajouter les lots de poissons capturés à la passerelle}
\usage{
mef_ajouter_lots(df)
}
\arguments{
\item{df}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
df doit contenir une variable "pre_id" (identifiant du prélèvement élémentaire).}
}
\value{
La passerelle complétée.
}
\description{
Ajouter les lots de poissons capturés à la passerelle
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_lots()
}
}
