% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_metriques.R
\name{mef_ajouter_metriques}
\alias{mef_ajouter_metriques}
\title{Ajouter à une passerelle les métriques IPR observées, théoriques et leur contribution
à l'indice IPR agrégé}
\usage{
mef_ajouter_metriques(df)
}
\arguments{
\item{df}{Dataframe qui doit comprendre une colonnes "ope_id" (identifiant de l'opération de pêche).}
}
\value{
Le dataframe passerelle complété par les métriques.
}
\description{
Ajouter à une passerelle les métriques IPR observées, théoriques et leur contribution
à l'indice IPR agrégé
}
\examples{
\dontrun{
passerelle_avec_metriques <- mef_ajouter_metriques(df = ma_passerelle)
}
}
