% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_passage.R
\name{mef_ajouter_passage}
\alias{mef_ajouter_passage}
\title{Ajouter à la passerelle, quand le prélèvement élémentaire est un passage, le numéro de passage}
\usage{
mef_ajouter_passage(df)
}
\arguments{
\item{df}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
df doit contenir une variable "pre_id" (identifiant du prélèvement élémentaire).}
}
\value{
La passerelle complétée.
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Le type de prélèvement doit donc être indiqué dans le data frame (Groupe de points,
Passage ou Ambiance). On applique mef_ajouter_passage() après mef_ajouter_type_prelevement().
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_type_prelevement() \%>\%
mef_ajouter_passage()
}
}
