% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_qualification.R
\name{mef_ajouter_qualification}
\alias{mef_ajouter_qualification}
\title{Ajouter à la passerelle la qualification de la donnée}
\usage{
mef_ajouter_qualification(df)
}
\arguments{
\item{df}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.
df doit contenir une variable "ope_id" (identifiant de l'opération de pêche).}
}
\value{
La passerelle complétée.
}
\description{
ex : Correcte / Incorrecte / Incertaine.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_qualification()
}
}
