% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_type_protocole.R
\name{mef_ajouter_type_protocole}
\alias{mef_ajouter_type_protocole}
\title{Ajouter le type de protocole à la passerelle}
\usage{
mef_ajouter_type_protocole(df)
}
\arguments{
\item{df}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
ex : Indice Abondance Saumon.
Nécessite d'avoir chargé les tables "operation" et "ref_protocole".
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_type_protocole()
}
}
