% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_validation.R
\name{mef_ajouter_validation}
\alias{mef_ajouter_validation}
\title{Ajouter à la passerelle l'a qualification de a donnée'état d'avancement de la saisie
    ex : En cours de saisie / Validé niveau 1.}
\usage{
mef_ajouter_validation(passerelle)
}
\arguments{
\item{passerelle}{Dataframe "passerelle" mettant en correspondance les identifiants des différentes tables.}
}
\value{
La passerelle complétée.
}
\description{
Ajouter à la passerelle l'a qualification de a donnée'état d'avancement de la saisie
    ex : En cours de saisie / Validé niveau 1.
}
\examples{
\dontrun{
passerelle <- passerelle \%>\%
mef_ajouter_validation()
}
}
