% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_nb_mini_annees.R
\name{mef_filtrer_nb_mini_annees}
\alias{mef_filtrer_nb_mini_annees}
\title{Filtrer un dataframe pour ne conserver que les points présents un nombre
    minimum d'années.}
\usage{
mef_filtrer_nb_mini_annees(df, nb_mini_annees)
}
\arguments{
\item{df}{Dataframe qui doit contenir le champ "pop_id".}

\item{nb_mini_annees}{Nombre entier indiquant le seuil de sélection (nb mini d'années
de données sur le point). Par défaut sa valeur est 1 donc tous les points avec
au moins une donnée sont sélectionnées.}
}
\value{
Dataframe filtré, c'est-à-dire exourgé des points qui n'ont pas assez d'années
    de données.
}
\description{
Filtrer un dataframe pour ne conserver que les points présents un nombre
    minimum d'années.
}
\examples{
\dontrun{
data_22_filtre <- mef_filtrer_nb_mini_annees(df = data_22, nb_mini_annees = 1)
}
}
