% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_ope_id.R
\name{mef_filtrer_ope_id}
\alias{mef_filtrer_ope_id}
\title{Filtrer les opérations de pêche}
\usage{
mef_filtrer_ope_id(
  df,
  var_id,
  var_tmp,
  var_surf = NULL,
  var_pro = NULL,
  var_pds = NULL,
  var_mei = NULL,
  var_pas = NULL,
  var_date = NULL,
  default = FALSE
)
}
\arguments{
\item{df}{Dataframe devant contenir le champ "ope_id"}

\item{var_id}{Nom(s) de la ou des variable(s) contenant les identifiants des points de prélèvement (ex : sta_id, pop_id, c(pop_id, saison)}

\item{var_tmp}{Nom de la variable contenant les dates de prélèvement (ex: annee)}

\item{var_surf}{[optionnelle] Nom de la variable contenant les surfaces de pêche : sélection des opérations dont la surface de pêche est renseignée}

\item{var_pro}{[optionnelle] Nom de la variable contenant les protocoles de pêche : sélection des opérations minimisant le nombre de changements de protocoles de pêche}

\item{var_pds}{[optionnelle] Nom de la variable contenant les poids : sélection des opérations dont le poids est renseignée}

\item{var_mei}{[optionnelle] Nom de la variable contenant des identifiants sur les mesures individuelles (ex : mei_id) : sélection des opérations associées à des mesures individuelles}

\item{var_pas}{[optionnelle] Nom de la variable contenant les numéros de passage de pêche : sélection des opérations associées à plusieurs passages de pêche}

\item{var_date}{[optionnelle] Nom de la variable contenant les dates additionnelles de pêche (ex : mois, saison) : sélection des opérations fournissant les séries les plus homogènes en terme de date}

\item{default}{Logique : si TRUE, selection de la première opération de pêche)}
}
\value{
Le dataframe filtré (ne contenant que les opérations de pêche sélectionnées).  
}
\description{
Lorsque plusieurs opérations de pêche par points de prélèvement et par date de prélèvement ont eu lieu, filtre les opérations de pêche en fonction des informations renseignées.
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
mef_ajouter_ope_date() \%>\%
mef_ajouter_surf_calc() \%>\%
mef_ajouter_type_protocole() \%>\%
mef_filtrer_ope_id(var_id = pop_id,
var_tmp = annee,
var_surf = ope_surface_calculee,
var_pro = pro_libelle,
default=TRUE)

df <- mef_creer_passerelle() \%>\%
mef_ajouter_ope_date() \%>\%
mef_ajouter_surf_calc() \%>\%
mef_ajouter_type_protocole() \%>\%
mef_ajouter_poids() \%>\%
mef_filtrer_ope_id(var_id = pop_id,
var_tmp = annee,
var_surf = ope_surface_calculee,
var_pds = lop_poids)
df <- df \%>\%
mef_filtrer_ope_id(var_id = pop_id,
var_tmp = annee,
var_pro = pro_libelle,
default=TRUE)
}
}
