% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtp_seuils.R
\name{qtp_seuils}
\alias{qtp_seuils}
\title{Calculer les seuils}
\usage{
qtp_seuils(df, code_espece, variable = "mei_taille", seuil_densite = 0.01)
}
\arguments{
\item{df}{Dataframe contenant les données.}

\item{code_espece}{Texte. Code espèce en trois lettres.}

\item{variable}{Texte. Variable, au choix "mei_taille" ou "mei_poids".}

\item{seuil_densite}{Numérique. Seuil de densité de probabilité probabilité au sens
de ggplot2::geom_density(). Par défaut il est de 1\%.}
}
\value{
Un vecteur contenant les seuils haut et bas, c'est-à-dire les tailles ou
    poids mini et maxi à retenir au seuil choisi.
}
\description{
Calculer les seuils
}
\examples{
\dontrun{
qtp_seuils (df = tp_data,
code_espece = "ROT",
variable = "mei_taille")
}
}
